fix(quotes): run-6 critique fixes + QuoteEditor module split
Defect fixes (critique run 6, score 31/40, canvas direction validated):

- Selected-lines deposit no longer a dead end: choosing the mode with no
  eligible line STAGES it locally (checkboxes + inline hint render, no
  doomed PATCH) and auto-persists once the first line is flagged; with an
  eligible line already present it saves immediately as before. The old
  flow PATCHed instantly, got rejected, snapped the select back, and
  unmounted the very checkboxes the error told the user to use.
- Send does the prerequisite instead of sitting disabled: the header Send
  stays enabled while edits settle - the click blurs the dirty field
  (starting its save) and queues the composer to open at quiescence, so
  the money moment never starts with a dead click. In the composer, a
  Send click with no valid recipients focuses the To field with a visible
  + aria-described reason; placeholder changed from an example address
  (which read as a filled value) to "Add recipient...".
- One canonical money story: list column header back to "Total" with an
  explicit "/first period" qualifier on recurring quotes (the list payload
  only carries the first-period figure); same qualifier in both bulk
  manifests; Detail suppresses $0.00 cadence rows like every other surface.
- Per-line image actions (add / from URL / remove) folded into the line's
  overflow menu - the last always-visible chrome on the canvas rows; only
  an upload spinner remains inline.
- Fixed-position line/block menus flip above the trigger near the viewport
  bottom; the totals live region no longer announces on mount (changes
  only); Stripe note copy disambiguated; token drift fixed (internal-band
  hsl literals -> foreground/70, Stripe warning -> warning tokens).

Mechanical split (soft 500-line guideline; was 4,322 lines):
- QuoteEditor.tsx (hub, ~2,200) + QuoteLineRows.tsx (GhostRow/Editable/
  Readonly rows + thumbs), QuoteBlockCard.tsx (+QuoteImagePreview),
  QuoteContractBlockEditor.tsx, quoteEditorShared.tsx (save-language
  plumbing: UNAUTHORIZED, LineUpdate, useSavedFlash, SrSaved, fieldRing,
  seamless). No behavioral changes from the split; tests untouched by it.

New/updated strings in all five locales. Web 412 tests green, astro check
clean; every fix verified live on the wt-stack (staged deposit incl. the
legitimate deposit-equals-total rejection path, one-click Send flush into
the composer, empty-To reason + focus, list qualifiers, image menu).
