+++ apps/web/src/components/billing/quotes/QuoteEditor.tsx
@@ -1,13 +1,11 @@
 import { Fragment, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Eye, EyeOff, GripVertical, Loader2, MoreHorizontal } from 'lucide-react';
+import { Eye, EyeOff, GripVertical, MoreHorizontal } from 'lucide-react';
 import '../../../lib/i18n';
-import { navigateTo } from '@/lib/navigation';
 import { fetchWithAuth } from '../../../stores/auth';
 import { runAction, handleActionError, ActionError } from '../../../lib/runAction';
 import { usePermissions } from '../../../lib/permissions';
 import { useOrgStore } from '../../../stores/orgStore';
-import { formatPercent } from '@/lib/i18n/format';
 import {
   addBlock,
   updateBlock,
@@ -21,7 +19,6 @@ import {
   reorderLines as reorderLinesApi,
   uploadQuoteImage,
   addQuoteImageFromUrl,
-  quoteImageUrl,
   updateQuote,
 } from '../../../lib/api/quotes';
 import {
@@ -32,34 +29,23 @@ import {
   type TemplateVersionSummary,
 } from '../../../lib/api/contractTemplates';
 import type { QuoteBlockInput, CoverPage } from '@breeze/shared';
-import { computeQuoteTotals, computeQuoteProfit, computeLineTotal, markupPct, priceFromMarkup, toCents, fromCents, toQuoteDepositConfig, type QuoteLineForMath, type QuoteProfit, type QuoteTotals, type QuoteDepositType, type QuoteDepositConfig } from '@breeze/shared';
-import { listCatalog, createCatalogItem, catalogItemImagePath, type CatalogItem } from '../../../lib/api/catalog';
+import { computeQuoteTotals, computeQuoteProfit, toQuoteDepositConfig, type QuoteLineForMath, type QuoteProfit, type QuoteTotals, type QuoteDepositType, type QuoteDepositConfig } from '@breeze/shared';
+import { listCatalog, createCatalogItem, type CatalogItem } from '../../../lib/api/catalog';
 import { ecExpressStatus, ecExpressImport, type EcProduct, type EcStatus, pax8Status, pax8Import, type Pax8Product, type Pax8PriceOption } from '../../../lib/api/distributors';
-import RichTextEditor from '../../common/RichTextEditor';
-import CatalogItemPicker from '../../catalog/CatalogItemPicker';
-import CatalogEnrichButton from '../../catalog/CatalogEnrichButton';
-import PolishButton from '../../catalog/PolishButton';
-import DistributorLookup from './DistributorLookup';
-import Pax8ProductLookup from './Pax8ProductLookup';
 import { ConfirmDialog } from '../../shared/ConfirmDialog';
+import { BlockCard, QuoteImagePreview } from './QuoteBlockCard';
+import { UNAUTHORIZED, type LineUpdate, SrSaved, fieldRing, useSavedFlash } from './quoteEditorShared';
 import { useMenuKeyboard } from '../shared/menuKeyboard';
 import { UnsavedBadge, RecurringBillingNote, MarginPanel } from '../billingUi';
-import { useAuthedImage } from './useQuoteImage';
 import {
   type QuoteDetail as QuoteDetailData,
   type QuoteBlock,
   type QuoteLine,
   type QuoteLineRecurrence,
-  type ContractBlockContent,
   formatMoney,
-  formatQuantity,
   pctFromFraction,
-  lineTaxAmount,
   lineTitle,
-  lineBlurb,
 } from './quoteTypes';
-
-const UNAUTHORIZED = () => void navigateTo('/login', { replace: true });
 
 // Phase 2: the add-block menu now offers `image` as well. An image block is
 // created with its uploaded `imageId` already in `content` — the editor uploads
@@ -75,13 +61,6 @@ const ADD_BLOCK_OPTIONS: { value: AddableBlockType; labelKey: string }[] = [
   { value: 'contract', labelKey: 'quotes.editor.blockTypes.contract' },
 ];
 
-const BLOCK_TYPE_LABEL_KEYS: Record<string, string> = {
-  heading: 'quotes.editor.blockTypes.heading',
-  rich_text: 'quotes.editor.blockTypes.richText',
-  image: 'quotes.editor.blockTypes.image',
-  line_items: 'quotes.editor.blockTypes.pricingTable',
-  contract: 'quotes.editor.blockTypes.contract',
-};
 
 /** Latest PUBLISHED version of a template (design: attach pins the latest
  *  published version, never a newer draft). Returns null when the template has
@@ -99,23 +78,6 @@ function unresolvedNamesFromMessage(message: string, knownNames: string[]): stri
   return knownNames.filter((name) => message.includes(name));
 }
 
-// Changed-fields payload for an inline line edit. Subset of
-// updateQuoteLineSchema (description/quantity/unitPrice/taxable/recurrence) —
-// the only fields the inline editor exposes.
-type LineUpdate = Partial<{
-  name: string | null;
-  description: string | null;
-  quantity: number;
-  unitPrice: number;
-  taxable: boolean;
-  recurrence: QuoteLineRecurrence;
-  unitCost: number | null;
-  sku: string | null;
-  partNumber: string | null;
-  imageId: string | null;
-  depositEligible: boolean;
-}>;
-
 interface Props {
   detail: QuoteDetailData;
   onChanged: () => void;
@@ -126,57 +88,6 @@ interface Props {
   onPendingEditsChange?: (hasPendingEdits: boolean) => void;
 }
 
-// Per-field blur-saves are confirmed by the amber dirty-ring clearing (sighted)
-// plus the SrSaved live region (screen readers) — NOT a toast. Toasts are
-// reserved for action-level events the user can't otherwise see (Line added,
-// Section removed, Proposal sent, Draft deleted), which fire their own
-// runAction successMessage. Per-field toasts were a storm during editing and
-// double-announced alongside SrSaved, so they were removed.
-
-// A transient "Saved" cue for the right-rail blur-to-save fields (terms, tax).
-// BlockCard and EditableLineRow replicate this same pattern inline rather than
-// calling the hook. Returns the on-flag (drives the SR live region) and a
-// trigger; clears its timer on unmount so a late fire can't setState a gone node.
-function useSavedFlash(): [boolean, () => void] {
-  const [on, setOn] = useState(false);
-  const timer = useRef<ReturnType<typeof setTimeout> | null>(null);
-  useEffect(() => () => { if (timer.current) clearTimeout(timer.current); }, []);
-  const flash = useCallback(() => {
-    setOn(true);
-    if (timer.current) clearTimeout(timer.current);
-    timer.current = setTimeout(() => setOn(false), 1500);
-  }, []);
-  return [on, flash];
-}
-
-// Visually-hidden polite live region — announces a transient "Saved" to screen
-// readers without taking visual space, pairing with the dirty-ring clearing that
-// sighted users see. The single per-field announcer (no toast), so SR users hear
-// "Saved" once, not twice. testId lets tests assert the cue fired.
-function SrSaved({ show, label, testId }: { show: boolean; label?: string; testId?: string }) {
-  const { t } = useTranslation('billing');
-  // role="status" already implies aria-live="polite" — don't double it.
-  return <span role="status" className="sr-only" data-testid={testId}>{show ? (label ?? t('quotes.editor.status.saved')) : ''}</span>;
-}
-
-// A field's save-state signal: amber BORDER while the edit is unsaved, a brief
-// green border pulse when it lands, nothing at rest. Border-color (not a ring):
-// the focus ring occupies the box-shadow channel, so a ring-based dirty signal
-// was painted over by focus on exactly the field being edited — the one moment
-// the signal matters. Border-color composes with the focus ring, never reflows
-// (the border is always present, only its color changes), and uses the
-// warning-strong indicator token (>=3:1 non-text on a light card).
-function fieldRing(dirty: boolean, saved: boolean): string {
-  return dirty ? 'border-warning-strong' : saved ? 'border-success' : '';
-}
-
-// Seamless (document-styled) field border: at rest the border is invisible so
-// the value reads as document text; hover/focus reveal the field. A state color
-// (dirty amber / saved green / error red) REPLACES the base set rather than
-// stacking on it, so two border-color utilities never compete.
-function seamless(state: string): string {
-  return state || 'border-transparent hover:border-border focus:border-border';
-}
 
 export default function QuoteEditor({ detail, onChanged, onPendingEditsChange }: Props) {
   const { t } = useTranslation('billing');
@@ -278,7 +189,7 @@ export default function QuoteEditor({ detail, onChanged, onPendingEditsChange }:
   // Per-block arrangement menu (grip + kebab live in the canvas margin, not in
   // a per-block header bar). One open menu at a time; same keyboard grammar as
   // every other menu (useMenuKeyboard).
-  const [blockMenu, setBlockMenu] = useState<{ id: string; top: number; left: number } | null>(null);
+  const [blockMenu, setBlockMenu] = useState<{ id: string; top: number; left: number; flip?: boolean } | null>(null);
   const blockMenuWrapRef = useRef<HTMLDivElement>(null);
   const blockMenuTriggerRef = useRef<HTMLButtonElement>(null);
   const { listRef: blockMenuListRef, onKeyDown: onBlockMenuKeyDown } = useMenuKeyboard(blockMenu !== null, () => setBlockMenu(null));
@@ -334,7 +245,10 @@ export default function QuoteEditor({ detail, onChanged, onPendingEditsChange }:
   // reverts itself. Cleared on the next keystroke. Server-side DEPOSIT_*
   // rejections (business rules the client can't know) still toast + resync.
   const [depositPctError, setDepositPctError] = useState<string | null>(null);
-  useEffect(() => { setDepositType(quote.depositType ?? 'none'); }, [quote.depositType]);
+  // Declared here (before the resync effect below) — see the staged
+  // selected_lines block further down for the rationale.
+  const stagedSelectedLines = useRef(false);
+  useEffect(() => { if (!stagedSelectedLines.current) setDepositType(quote.depositType ?? 'none'); }, [quote.depositType]);
   useEffect(() => { setDepositPercentDraft(quote.depositPercent ?? ''); }, [quote.depositPercent]);
 
   // Coalesce re-pulls: each mutation calls refresh(), but tab-through editing
@@ -495,9 +409,32 @@ export default function QuoteEditor({ detail, onChanged, onPendingEditsChange }:
     return { pct, error: null };
   }, [t]);
 
+  // 'selected_lines' with no eligible lines yet is a STAGED state, not an
+  // error: PATCHing immediately made the server reject, the select snap back,
+  // and the eligibility checkboxes (which only render in this mode) unmount —
+  // the error instructed an action the UI had just made impossible. Instead the
+  // mode holds locally (checkboxes + inline hint visible) and persists itself
+  // the moment the first line is flagged. The ref guards the server-resync
+  // effect from clobbering the staged mode on unrelated refreshes.
+  const hasEligibleLines = lines.some((l) => l.depositEligible);
+  useEffect(() => {
+    if (!stagedSelectedLines.current) return;
+    if (!hasEligibleLines) return;
+    stagedSelectedLines.current = false;
+    void saveDeposit({ depositType: 'selected_lines' }).then((ok) => { if (!ok) revertDepositMirrors(); });
+    // saveDeposit/revertDepositMirrors are stable useCallbacks from above.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [hasEligibleLines]);
+
   const onDepositTypeChange = useCallback((next: QuoteDepositType) => {
     setDepositType(next);
+    stagedSelectedLines.current = false;
     if (next !== 'percent') setDepositPctError(null);
+    if (next === 'selected_lines' && !lines.some((l) => l.depositEligible)) {
+      // Stage: show the per-line checkboxes + hint; persist on first flag.
+      stagedSelectedLines.current = true;
+      return;
+    }
     if (next === 'percent') {
       // Saving type='percent' with a null percent would 400 DEPOSIT_PERCENT_INVALID,
       // so defer the PATCH until a valid percent exists — persist immediately only
@@ -731,7 +668,12 @@ export default function QuoteEditor({ detail, onChanged, onPendingEditsChange }:
   const SR_SETTLE_MS = 800;
   const [srAnnouncement, setSrAnnouncement] = useState('');
   const srTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
+  // Skip the very first sentence: without this the debounce fires ~800ms after
+  // mount and announces the initial totals to an SR user who hasn't edited
+  // anything yet. Only CHANGES announce.
+  const srMounted = useRef(false);
   useEffect(() => {
+    if (!srMounted.current) { srMounted.current = true; return; }
     if (srTimer.current) clearTimeout(srTimer.current);
     srTimer.current = setTimeout(() => setSrAnnouncement(srSentence), SR_SETTLE_MS);
     return () => { if (srTimer.current) clearTimeout(srTimer.current); };
@@ -1936,7 +1878,8 @@ export default function QuoteEditor({ detail, onChanged, onPendingEditsChange }:
                           onClick={(e) => {
                             if (blockMenu?.id === block.id) { setBlockMenu(null); return; }
                             const r = e.currentTarget.getBoundingClientRect();
-                            setBlockMenu({ id: block.id, top: r.bottom + 4, left: r.left });
+                            const flip = r.bottom + 160 > window.innerHeight;
+                            setBlockMenu({ id: block.id, top: flip ? r.top - 4 : r.bottom + 4, left: r.left, flip });
                           }}
                           aria-haspopup="menu"
                           aria-expanded={blockMenu?.id === block.id}
@@ -1955,7 +1898,7 @@ export default function QuoteEditor({ detail, onChanged, onPendingEditsChange }:
                           ref={blockMenuListRef}
                           onKeyDown={onBlockMenuKeyDown}
                           aria-label={t('quotes.editor.actions.sectionActions')}
-                          style={{ position: 'fixed', top: blockMenu.top, left: blockMenu.left }}
+                          style={{ position: 'fixed', top: blockMenu.top, left: blockMenu.left, transform: blockMenu.flip ? 'translateY(-100%)' : undefined }}
                           className="z-50 w-max min-w-40 rounded-md border bg-card py-1 shadow-md"
                           data-testid={`quote-block-actions-menu-${block.id}`}
                         >
@@ -2162,8 +2105,13 @@ export default function QuoteEditor({ detail, onChanged, onPendingEditsChange }:
                   </>
                 )}
                 {depositType === 'selected_lines' && (
-                  <p className="text-xs text-muted-foreground">
-                    {t('quotes.editor.deposit.selectedLinesHelp')}
+                  <p
+                    className={quote.depositType !== 'selected_lines' ? 'text-xs font-medium text-warning-foreground dark:text-warning' : 'text-xs text-muted-foreground'}
+                    data-testid="quote-deposit-selected-hint"
+                  >
+                    {quote.depositType !== 'selected_lines'
+                      ? t('quotes.editor.deposit.selectedLinesStagedHint')
+                      : t('quotes.editor.deposit.selectedLinesHelp')}
                   </p>
                 )}
               </div>
@@ -2357,1913 +2305,4 @@ function InsertGap({ index, active, onToggle, label, dropActive, onDragOver, onD
       </button>
     </div>
   );
-}
-
-// The ghost row: an always-ready entry row at the foot of every pricing table.
-// Type a name, Tab through qty/price/cadence, press Enter — the line commits
-// and focus returns to the name for the next one. This is the fast lane for the
-// daily compose loop; the full picker (catalog / AI lookup / distributor / SKU
-// and cost fields) stays available behind the "More ways to add" disclosure.
-function GhostRow({ blockId, busy, onAdd, colSpan }: {
-  blockId: string;
-  busy: boolean;
-  onAdd: (form: { name: string; description: string; quantity: string; unitPrice: string; cost: string; sku: string; partNumber: string; taxable: boolean; recurrence: QuoteLineRecurrence; saveToCatalog: boolean }) => Promise<boolean>;
-  colSpan: number;
-}) {
-  const { t } = useTranslation('billing');
-  const [name, setName] = useState('');
-  const [qty, setQty] = useState('1');
-  const [price, setPrice] = useState('');
-  const [rec, setRec] = useState<QuoteLineRecurrence>('one_time');
-  const [error, setError] = useState<string | null>(null);
-  const nameRef = useRef<HTMLInputElement>(null);
-
-  const commit = async () => {
-    if (!name.trim()) return; // nothing typed — Enter is a no-op, not an error
-    const qtyNum = Number(qty);
-    if (!Number.isFinite(qtyNum) || qtyNum <= 0 || !Number.isInteger(qtyNum)) {
-      setError(t('quotes.editor.errors.quantityWholeGreaterThanZero'));
-      return;
-    }
-    const priceNum = price.trim() === '' ? 0 : Number(price);
-    if (!Number.isFinite(priceNum) || priceNum < 0) {
-      setError(t('quotes.editor.errors.unitPriceZeroOrMore'));
-      return;
-    }
-    setError(null);
-    const ok = await onAdd({
-      name, description: '', quantity: qty, unitPrice: price.trim() === '' ? '0' : price,
-      cost: '', sku: '', partNumber: '', taxable: false, recurrence: rec, saveToCatalog: false,
-    });
-    if (ok) {
-      setName(''); setQty('1'); setPrice(''); setRec('one_time');
-      // Refocus after React re-enables the input: the field is disabled during
-      // the save (which also DROPS focus — disabling the focused element moves
-      // focus to <body>), and the busy=false re-render lands async. Retry over
-      // a few frames until the node is focusable again.
-      const refocus = (tries: number) => {
-        const el = nameRef.current;
-        if (el && !el.disabled) { el.focus(); return; }
-        if (tries > 0) requestAnimationFrame(() => refocus(tries - 1));
-      };
-      requestAnimationFrame(() => refocus(10));
-    }
-  };
-  const onKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === 'Enter') { e.preventDefault(); void commit(); }
-  };
-  const ghostField = 'h-9 rounded-md border border-transparent bg-transparent transition-colors hover:border-border focus:border-border focus:outline-hidden disabled:opacity-60';
-
-  return (
-    <>
-      <tr className="border-t" data-testid={`quote-ghost-row-${blockId}`}>
-        <td className="px-1.5 py-1.5">
-          <input
-            ref={nameRef}
-            type="text"
-            value={name}
-            onChange={(e) => { setName(e.target.value); setError(null); }}
-            onKeyDown={onKeyDown}
-            disabled={busy}
-            placeholder={t('quotes.editor.ghost.placeholder')}
-            aria-label={t('quotes.editor.ghost.nameAria')}
-            data-testid={`quote-ghost-name-${blockId}`}
-            className={`${ghostField} w-full px-2 text-sm placeholder:text-muted-foreground/70`}
-          />
-        </td>
-        <td className="px-1.5 py-1.5 text-right">
-          <input
-            type="number" min="1" step="1"
-            value={qty}
-            onChange={(e) => { setQty(e.target.value); setError(null); }}
-            onKeyDown={onKeyDown}
-            disabled={busy}
-            aria-label={t('quotes.editor.line.quantityAria')}
-            data-testid={`quote-ghost-qty-${blockId}`}
-            className={`${ghostField} w-14 px-2 text-right text-sm tabular-nums`}
-          />
-        </td>
-        <td className="px-1.5 py-1.5 text-right">
-          <input
-            type="number" min="0" step="0.01"
-            value={price}
-            onChange={(e) => { setPrice(e.target.value); setError(null); }}
-            onKeyDown={onKeyDown}
-            disabled={busy}
-            placeholder="0.00"
-            aria-label={t('quotes.editor.table.unitPrice')}
-            data-testid={`quote-ghost-price-${blockId}`}
-            className={`${ghostField} w-24 px-2 text-right text-sm tabular-nums`}
-          />
-          <select
-            value={rec}
-            onChange={(e) => setRec(e.target.value as QuoteLineRecurrence)}
-            onKeyDown={onKeyDown}
-            disabled={busy}
-            aria-label={t('quotes.editor.line.billingFrequencyAria')}
-            data-testid={`quote-ghost-recurrence-${blockId}`}
-            className="ml-auto mt-1 block h-7 w-24 rounded-md border border-transparent bg-transparent py-0 pl-2 pr-6 text-xs text-muted-foreground transition-colors hover:border-border focus:border-border focus:outline-hidden disabled:opacity-60"
-          >
-            <option value="one_time">{t('quotes.editor.recurrence.one_time')}</option>
-            <option value="monthly">{t('quotes.editor.recurrence.monthly')}</option>
-            <option value="annual">{t('quotes.editor.recurrence.annual')}</option>
-          </select>
-        </td>
-        <td className="px-1.5 py-1.5 text-right align-middle">
-          {name.trim() && (
-            <button
-              type="button"
-              onClick={() => void commit()}
-              disabled={busy}
-              data-testid={`quote-ghost-add-${blockId}`}
-              className="inline-flex h-7 items-center gap-1 rounded-md border px-2 text-xs font-medium hover:bg-muted disabled:opacity-50"
-            >
-              {busy && <Loader2 className="h-3 w-3 animate-spin" aria-hidden="true" />}
-              {t('quotes.editor.ghost.add')}
-            </button>
-          )}
-        </td>
-        <td className="px-1.5 py-1.5" />
-      </tr>
-      {error && (
-        <tr className="border-0">
-          <td colSpan={colSpan} className="px-1.5 pb-1.5">
-            <p className="text-xs text-destructive" data-testid={`quote-ghost-error-${blockId}`}>{error}</p>
-          </td>
-        </tr>
-      )}
-    </>
-  );
-}
-
-// ── A single block, with an inline line builder when it is a pricing table ──
-function BlockCard({
-  block, quoteId, lines, currency, taxRate, catalog, catalogLoadFailed, isPending, canWrite, showInternal, depositSelectMode, ecActive, pax8Active, defaultMarkupPct, onAddCatalog, onImportAddDistributor, onImportAddPax8, onAddManual, onEditLine, onEditBlock, onMoveLine, onRemoveLine, onLineDraft,
-  moveTargets, onMoveLineToBlock,
-}: {
-  block: QuoteBlock;
-  quoteId: string;
-  lines: QuoteLine[];
-  currency: string;
-  taxRate: string | null;
-  catalog: CatalogItem[];
-  catalogLoadFailed: boolean;
-  isPending: (key: string) => boolean;
-  canWrite: boolean;
-  showInternal: boolean;
-  /** When true (quote deposit = 'selected_lines'), each editable line row shows a
-   *  deposit-eligible checkbox. */
-  depositSelectMode: boolean;
-  ecActive: boolean;
-  pax8Active: boolean;
-  /** Partner default markup % for pre-pricing AI auto-filled lines; null = unknown. */
-  defaultMarkupPct: number | null;
-  onAddCatalog: (blockId: string, item: CatalogItem) => void;
-  onImportAddDistributor: (blockId: string, product: EcProduct, sellPrice: number) => void;
-  onImportAddPax8: (blockId: string, product: Pax8Product, term: Pax8PriceOption, sellPrice: number) => void;
-  onAddManual: (
-    blockId: string,
-    form: { name: string; description: string; quantity: string; unitPrice: string; cost: string; sku: string; partNumber: string; taxable: boolean; recurrence: QuoteLineRecurrence; saveToCatalog: boolean },
-  ) => Promise<boolean>;
-  onEditLine: (lineId: string, body: LineUpdate, scopeKey?: string) => Promise<boolean>;
-  onEditBlock: (block: QuoteBlock, content: Record<string, unknown>) => Promise<boolean>;
-  onMoveLine: (line: QuoteLine, direction: 'up' | 'down') => void;
-  onRemoveLine: (line: QuoteLine) => void;
-  onLineDraft: (lineId: string, draft: QuoteLineForMath | null) => void;
-  /** Other pricing panels this block's lines can move to (empty → control hidden). */
-  moveTargets: { id: string; label: string }[];
-  onMoveLineToBlock: (line: QuoteLine, targetBlockId: string) => void;
-}) {
-  const { t } = useTranslation('billing');
-  // Pending state scoped to this block: editing/removing this block, or adding a
-  // line to it, never disables anything in a sibling block.
-  const blockBusy = isPending(`block:${block.id}`);
-  const addLineBusy = isPending(`add-line:${block.id}`);
-
-  const [pickerOpen, setPickerOpen] = useState(false);
-  const [mode, setMode] = useState<'catalog' | 'manual' | 'distributor' | 'pax8'>('catalog');
-  const [name, setName] = useState('');
-  const [desc, setDesc] = useState('');
-  const [qty, setQty] = useState('1');
-  const [price, setPrice] = useState('0.00');
-  const [cost, setCost] = useState('');
-  const [markup, setMarkup] = useState('');
-  const [sku, setSku] = useState('');
-  const [partNumber, setPartNumber] = useState('');
-  const [taxable, setTaxable] = useState(false);
-  const [recurrence, setRecurrence] = useState<QuoteLineRecurrence>('one_time');
-  const [saveToCatalog, setSaveToCatalog] = useState(false);
-  // What the last auto-fill touched, for the "Auto-filled: …" summary line.
-  // Cleared when the form resets (successful add) or a new query starts.
-  const [autoFilled, setAutoFilled] = useState<string[] | null>(null);
-
-  // Two-way price ↔ markup% coupling (cost is always an input, never derived).
-  // Whichever of price/markup the user set last stays authoritative: editing it
-  // recomputes the other, and a later cost edit recomputes the derived one.
-  const priceAuthority = useRef<'price' | 'markup'>('price');
-  // Live mirrors for the enrich onApply callback: the web lookup takes seconds,
-  // so the closure's cost/price are stale by the time the result lands — the
-  // pristine-field checks must read the CURRENT values or auto-fill could
-  // overwrite a number the user typed mid-search.
-  const costRef = useRef(cost); costRef.current = cost;
-  const priceRef = useRef(price); priceRef.current = price;
-  const deriveMarkup = (nextPrice: string, nextCost: string) => {
-    // A zero/empty price is the form's pristine state, not a -100% pricing
-    // decision — show no markup rather than a misleading negative.
-    if (nextPrice.trim() === '' || Number(nextPrice) === 0) { setMarkup(''); return; }
-    const mk = markupPct(nextPrice, nextCost);
-    setMarkup(mk === null ? '' : String(Number(mk.toFixed(2))));
-  };
-  const derivePrice = (nextMarkup: string, nextCost: string) => {
-    const m = Number(nextMarkup);
-    if (nextCost.trim() === '' || nextMarkup.trim() === '' || !Number.isFinite(m) || Number(nextCost) <= 0) return;
-    setPrice(priceFromMarkup(nextCost, m));
-  };
-  const onPriceChange = (v: string) => {
-    setPrice(v);
-    priceAuthority.current = 'price';
-    deriveMarkup(v, cost);
-  };
-  const onMarkupChange = (v: string) => {
-    setMarkup(v);
-    priceAuthority.current = 'markup';
-    derivePrice(v, cost);
-  };
-  const onCostChange = (v: string) => {
-    setCost(v);
-    if (priceAuthority.current === 'markup') derivePrice(markup, v);
-    else deriveMarkup(price, v);
-  };
-
-  const isTable = block.blockType === 'line_items';
-  const heading = (block.content?.text as string | undefined) ?? '';
-  const html = (block.content?.html as string | undefined) ?? '';
-  const tableLabel = (block.content?.label as string | undefined) ?? '';
-  const showSubtotal = (block.content?.showSubtotal as boolean | undefined) === true;
-  const imageId = (block.content?.imageId as string | undefined) ?? '';
-  const imageCaption = (block.content?.caption as string | undefined) ?? '';
-
-  // Inline drafts for editable block content; resync if the persisted value
-  // changes (e.g. after a refresh) so server normalization wins.
-  const [headingDraft, setHeadingDraft] = useState(heading);
-  const [richDraft, setRichDraft] = useState(html);
-  const [labelDraft, setLabelDraft] = useState(tableLabel);
-  // Resync drafts from the server only when the user hasn't diverged from what we
-  // last showed. A quiet reload (fired by an unrelated inline edit elsewhere) must
-  // not clobber heading/rich text this user is mid-edit in: if the local draft no
-  // longer matches the prop we last synced, the user has typed — keep their text.
-  const lastHeading = useRef(heading);
-  const lastHtml = useRef(html);
-  const lastLabel = useRef(tableLabel);
-  // Set right after our own rich_text save so the next html-prop resync adopts
-  // the server-normalized body unconditionally (see commitRich).
-  const forceRichResync = useRef(false);
-  useEffect(() => {
-    setHeadingDraft((cur) => (cur === lastHeading.current ? heading : cur));
-    lastHeading.current = heading;
-  }, [heading]);
-  useEffect(() => {
-    setRichDraft((cur) => (forceRichResync.current || cur === lastHtml.current ? html : cur));
-    forceRichResync.current = false;
-    lastHtml.current = html;
-  }, [html]);
-  useEffect(() => {
-    setLabelDraft((cur) => (cur === lastLabel.current ? tableLabel : cur));
-    lastLabel.current = tableLabel;
-  }, [tableLabel]);
-
-  // Quiet "Saved" flash for inline content edits (replaces the old per-edit
-  // success toast). Cleared on unmount so a late timer can't setState a gone row.
-  const [blockSaved, setBlockSaved] = useState(false);
-  const savedTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
-  useEffect(() => () => { if (savedTimer.current) clearTimeout(savedTimer.current); }, []);
-  const flashSaved = useCallback(() => {
-    setBlockSaved(true);
-    if (savedTimer.current) clearTimeout(savedTimer.current);
-    savedTimer.current = setTimeout(() => setBlockSaved(false), 1500);
-  }, []);
-
-  const commitHeading = async () => {
-    const text = headingDraft.trim();
-    if (!text || text === heading) { setHeadingDraft(heading); return; }
-    if (await onEditBlock(block, { text, level: (block.content?.level as number | undefined) ?? 2 })) flashSaved();
-  };
-  const commitRich = async () => {
-    if (richDraft === html) return;
-    if (await onEditBlock(block, { html: richDraft })) {
-      // The server sanitizes the body on write (rel/attribute normalization), so
-      // the reloaded html prop is the source of truth. Force the next resync to
-      // adopt it — otherwise a residual TipTap-vs-sanitizer string mismatch would
-      // keep the block flagged "unsaved" and re-PATCH on every blur. Mirrors the
-      // force-reseed TemplateEditor uses after saveDraft.
-      forceRichResync.current = true;
-      flashSaved();
-    }
-  };
-  // Rename a pricing table. An empty label is a valid clear (the document falls
-  // back to its "Pricing" default), so — unlike heading — we commit the trimmed
-  // value even when blank, only skipping when nothing actually changed.
-  // Both the label and the subtotal toggle live in the SAME line_items content
-  // object, and onEditBlock REPLACES content wholesale — so each edit must carry
-  // the other's current value forward or it gets dropped.
-  const lineItemsContent = (nextLabel: string, nextSubtotal: boolean) => ({
-    ...(nextLabel.trim() ? { label: nextLabel.trim() } : {}),
-    ...(nextSubtotal ? { showSubtotal: true } : {}),
-  });
-  const commitLabel = async () => {
-    const label = labelDraft.trim();
-    if (label === tableLabel.trim()) { setLabelDraft(tableLabel); return; }
-    if (await onEditBlock(block, lineItemsContent(label, showSubtotal))) flashSaved();
-  };
-  const toggleSubtotal = async (next: boolean) => {
-    if (await onEditBlock(block, lineItemsContent(tableLabel, next))) flashSaved();
-  };
-
-  // Inline errors for the manual-line form's qty/price/cost — same contract as
-  // the edit-row fields (aria-invalid + destructive ring + text under the input).
-  // The parent's addManual gates stay as a backstop, but validating here means
-  // the message lands next to the field, not in a bottom-corner toast.
-  const [manualErrors, setManualErrors] = useState<{ qty?: string; price?: string; cost?: string }>({});
-  const clearManualError = (field: 'qty' | 'price' | 'cost') =>
-    setManualErrors((e) => { if (!(field in e)) return e; const n = { ...e }; delete n[field]; return n; });
-
-  const submitManual = async () => {
-    const errs: { qty?: string; price?: string; cost?: string } = {};
-    const qtyNum = Number(qty);
-    if (!Number.isFinite(qtyNum) || qtyNum <= 0 || !Number.isInteger(qtyNum)) errs.qty = t('quotes.editor.errors.quantityWholeGreaterThanZero');
-    const priceNum = Number(price);
-    if (!Number.isFinite(priceNum) || priceNum < 0) errs.price = t('quotes.editor.errors.unitPriceZeroOrMore');
-    if (cost.trim() !== '' && (!Number.isFinite(Number(cost)) || Number(cost) < 0)) errs.cost = t('quotes.editor.errors.costZeroOrMore');
-    setManualErrors(errs);
-    if (Object.keys(errs).length > 0) return;
-    const ok = await onAddManual(block.id, { name, description: desc, quantity: qty, unitPrice: price, cost, sku, partNumber, taxable, recurrence, saveToCatalog });
-    // Only clear the form on success, so a rejected add (e.g. qty 0) keeps the
-    // user's input to correct rather than wiping it.
-    if (ok) {
-      setName(''); setDesc(''); setQty('1'); setPrice('0.00'); setCost(''); setMarkup(''); setSku(''); setPartNumber('');
-      setTaxable(false); setRecurrence('one_time'); setSaveToCatalog(false); setAutoFilled(null);
-      priceAuthority.current = 'price';
-    }
-  };
-
-  return (
-    <section data-testid={`quote-block-${block.id}`}>
-      {/* No card chrome, no uppercase type-label bar: on the canvas a block IS
-          its content (a heading looks like a heading, a table like a table).
-          The type is self-evident; arrangement lives in the gutter controls. */}
-      <SrSaved show={blockSaved} testId={`quote-block-saved-${block.id}`} />
-      <div>
-        {block.blockType === 'heading' && (
-          canWrite ? (
-            <input
-              value={headingDraft}
-              aria-label={t('quotes.editor.addSection.headingPlaceholder')}
-              onChange={(e) => setHeadingDraft(e.target.value)}
-              onBlur={() => void commitHeading()}
-              disabled={blockBusy}
-              data-testid={`quote-block-heading-input-${block.id}`}
-              className={`w-full rounded-md border bg-transparent px-2 py-1 text-lg font-semibold transition-colors focus:outline-hidden disabled:opacity-60 ${seamless(fieldRing(headingDraft.trim() !== heading, blockSaved))}`}
-            />
-          ) : (
-            <p className="text-lg font-semibold" data-testid={`quote-block-heading-content-${block.id}`}>{heading}</p>
-          )
-        )}
-        {block.blockType === 'rich_text' && (
-          canWrite ? (
-            // The editor commits on blur (same as the old textarea). React's
-            // onBlur fires on focusout of the contenteditable; toolbar buttons
-            // preventDefault their mousedown so clicking them never blurs the
-            // editor and never triggers a spurious commit.
-            <div
-              onBlur={() => void commitRich()}
-              data-testid={`quote-block-rich-input-${block.id}`}
-              className={`rounded-md transition-shadow ${fieldRing(richDraft !== html, blockSaved)}`}
-            >
-              <RichTextEditor
-                value={richDraft}
-                onChange={setRichDraft}
-                ariaLabel={t('quotes.editor.block.richTextContentAria')}
-                testId={`quote-block-rich-editor-${block.id}`}
-              />
-            </div>
-          ) : (
-            // Read-only (no write permission): the API sanitizes every rich_text
-            // block to the fixed p/br/strong/em/u/h3/h4/ul/ol/li/a allowlist on
-            // read serialization (richTextSanitize.ts), so rendering it as real
-            // HTML here is safe — same pattern as QuoteDocument.
-            <div
-              className="quote-rich-text prose prose-sm max-w-none text-sm text-foreground dark:prose-invert"
-              data-testid={`quote-block-rich-content-${block.id}`}
-              dangerouslySetInnerHTML={{ __html: html }}
-            />
-          )
-        )}
-        {block.blockType === 'image' && (
-          imageId ? (
-            <figure className="space-y-1" data-testid={`quote-block-image-content-${block.id}`}>
-              <QuoteImagePreview quoteId={quoteId} imageId={imageId} caption={imageCaption} />
-              {imageCaption && <figcaption className="text-xs text-muted-foreground">{imageCaption}</figcaption>}
-            </figure>
-          ) : (
-            <p className="text-sm text-muted-foreground">{t('quotes.editor.block.imageSectionPdf')}</p>
-          )
-        )}
-
-        {block.blockType === 'contract' && (
-          <ContractBlockEditor block={block} canWrite={canWrite} onEditBlock={onEditBlock} />
-        )}
-
-        {isTable && (
-          <div className="space-y-3">
-            {canWrite && (
-              <input
-                type="text"
-                value={labelDraft}
-                aria-label={t('quotes.editor.table.labelAria')}
-                onChange={(e) => setLabelDraft(e.target.value)}
-                onBlur={() => void commitLabel()}
-                disabled={blockBusy}
-                placeholder={t('quotes.editor.table.labelPlaceholder')}
-                data-testid={`quote-block-table-label-input-${block.id}`}
-                className={`h-9 w-full rounded-md border bg-transparent px-2 text-sm font-semibold transition-colors focus:outline-hidden disabled:opacity-60 ${seamless(fieldRing(labelDraft.trim() !== tableLabel.trim(), blockSaved))}`}
-              />
-            )}
-            {canWrite && (
-              <label className="flex items-center gap-2 text-xs text-muted-foreground">
-                <input
-                  type="checkbox"
-                  checked={showSubtotal}
-                  onChange={(e) => void toggleSubtotal(e.target.checked)}
-                  disabled={blockBusy}
-                  data-testid={`quote-block-subtotal-toggle-${block.id}`}
-                />
-                {t('quotes.editor.table.showSubtotal')}
-              </label>
-            )}
-            {/* Four data columns (Item flexes, Qty/Price/Total are content-sized) so
-                the per-line Total — the most-checked figure on a quote — is always
-                visible without sideways scrolling at desktop widths. Billing cadence
-                rides in the Price cell; Taxable moved to each line's controls row;
-                per-line tax renders as a sub-line under the Total. The wrapper still
-                scrolls on genuinely narrow screens (phone), without a sticky column. */}
-            <div className="overflow-x-auto rounded-md focus:outline-hidden focus-visible:ring-2 focus-visible:ring-ring" role="region" aria-label={t('quotes.editor.table.scrollAria')} tabIndex={0}>
-            <table className="w-full min-w-[36rem] text-sm" data-testid={`quote-block-lines-${block.id}`}>
-              <thead>
-                <tr className="border-b text-left text-xs uppercase tracking-wide text-muted-foreground">
-                  <th className="min-w-[12rem] px-1.5 py-2 font-medium">{t('quotes.editor.table.item')}</th>
-                  <th className="px-1.5 py-2 text-right font-medium">{t('quotes.editor.table.qty')}</th>
-                  <th className="px-1.5 py-2 text-right font-medium">{t('quotes.editor.table.unitPrice')}</th>
-                  <th className="px-1.5 py-2 text-right font-medium">{t('quotes.editor.table.total')}</th>
-                  {canWrite && <th className="px-1.5 py-2" />}
-                </tr>
-              </thead>
-              <tbody>
-                {lines.length === 0 && !canWrite ? (
-                  <tr>
-                    <td colSpan={4} className="px-2 py-6 text-center text-sm text-muted-foreground">
-                      {t('quotes.editor.table.emptyLines')}
-                    </td>
-                  </tr>
-                ) : (
-                  lines.map((l, idx) =>
-                    canWrite ? (
-                      <EditableLineRow
-                        key={l.id}
-                        line={l}
-                        quoteId={quoteId}
-                        currency={currency}
-                        taxRate={taxRate}
-                        isPending={isPending}
-                        isFirst={idx === 0}
-                        isLast={idx === lines.length - 1}
-                        showInternal={showInternal}
-                        depositSelectMode={depositSelectMode}
-                        onEdit={onEditLine}
-                        onMove={onMoveLine}
-                        onRemove={onRemoveLine}
-                        onDraft={onLineDraft}
-                        moveTargets={moveTargets}
-                        onMoveTo={onMoveLineToBlock}
-                      />
-                    ) : (
-                      <ReadonlyLineRow key={l.id} line={l} quoteId={quoteId} currency={currency} taxRate={taxRate} isFirst={idx === 0} showInternal={showInternal} />
-                    ),
-                  )
-                )}
-                {/* Ghost row: the fast lane for manual entry — always ready at
-                    the table foot, Enter commits and refocuses for the next. */}
-                {canWrite && (
-                  <GhostRow
-                    blockId={block.id}
-                    busy={addLineBusy}
-                    onAdd={(form) => onAddManual(block.id, form)}
-                    colSpan={5}
-                  />
-                )}
-              </tbody>
-            </table>
-            </div>
-
-            {/* The full add-line picker (catalog / AI lookup / distributor /
-                SKU + cost fields) collapses behind a disclosure — the ghost row
-                covers the fast manual path, so this chrome only renders when a
-                tech asks for the heavier modes. */}
-            {canWrite && (
-              <button
-                type="button"
-                onClick={() => setPickerOpen((v) => !v)}
-                aria-expanded={pickerOpen}
-                data-testid={`quote-block-add-line-toggle-${block.id}`}
-                className="mt-2 inline-flex items-center gap-1 rounded-md border border-transparent px-2 py-1 text-xs font-medium text-muted-foreground hover:border-border hover:text-foreground"
-              >
-                <span aria-hidden="true">{pickerOpen ? '−' : '+'}</span> {t('quotes.editor.addLine.moreWays')}
-              </button>
-            )}
-            {canWrite && pickerOpen && (
-            <div className="mt-1 rounded-md border bg-background/40 p-4" data-testid={`quote-block-add-line-${block.id}`}>
-              <div className="mb-3 flex gap-2">
-                {(['catalog', 'manual', ...(ecActive ? ['distributor'] as const : []), ...(pax8Active ? ['pax8'] as const : [])] as const).map((m) => (
-                  <button
-                    key={m}
-                    type="button"
-                    aria-pressed={mode === m}
-                    onClick={() => setMode(m)}
-                    data-testid={`quote-line-mode-${block.id}-${m}`}
-                    className={`rounded-md border px-3 py-1 text-xs font-medium ${
-                      mode === m ? 'border-primary bg-primary/10 text-primary' : 'hover:bg-muted'
-                    }`}
-                  >
-                    {m === 'catalog'
-                      ? t('quotes.editor.addLine.catalogItem')
-                      : m === 'manual'
-                        ? t('quotes.editor.addLine.manualLine')
-                        : m === 'distributor'
-                          ? t('quotes.editor.addLine.searchDistributor')
-                          : t('quotes.editor.addLine.searchPax8')}
-                  </button>
-                ))}
-              </div>
-
-              {mode === 'distributor' ? (
-                <DistributorLookup
-                  blockId={block.id}
-                  busy={addLineBusy}
-                  onImportAdd={(product, sellPrice) => onImportAddDistributor(block.id, product, sellPrice)}
-                />
-              ) : mode === 'pax8' ? (
-                <Pax8ProductLookup
-                  blockId={block.id}
-                  busy={addLineBusy}
-                  onImportAdd={(product, term, sellPrice) => onImportAddPax8(block.id, product, term, sellPrice)}
-                />
-              ) : mode === 'catalog' ? (
-                catalog.length === 0 ? (
-                  catalogLoadFailed ? (
-                    <p className="text-xs text-muted-foreground" data-testid={`quote-catalog-error-${block.id}`}>
-                      {t('quotes.editor.catalog.loadError')}
-                    </p>
-                  ) : (
-                    <p className="text-xs text-muted-foreground" data-testid={`quote-catalog-empty-${block.id}`}>
-                      {t('quotes.editor.catalog.empty')}{' '}
-                      <a href="/settings/catalog" className="underline hover:text-foreground">{t('quotes.editor.catalog.addSome')}</a>.
-                    </p>
-                  )
-                ) : (
-                  <CatalogItemPicker
-                    items={catalog}
-                    includeBundles={false}
-                    onSelect={(it) => onAddCatalog(block.id, it)}
-                    testId={`quote-catalog-picker-${block.id}`}
-                    placeholder={t('quotes.editor.catalog.searchPlaceholder')}
-                    disabled={addLineBusy}
-                  />
-                )
-              ) : (
-                <div className="space-y-3">
-                  <div className="flex flex-wrap items-center gap-2">
-                    <CatalogEnrichButton
-                      idSuffix={`quote-${block.id}`}
-                      helpText={
-                        defaultMarkupPct != null
-                          ? t('quotes.editor.autoFill.helpWithDefaultMarkup', { markup: String(Number(defaultMarkupPct)) })
-                          : t('quotes.editor.autoFill.help')
-                      }
-                      guidanceSuffix={null}
-                      onApply={(result) => {
-                        const d = result.draft;
-                        setName(d.name);
-                        setDesc(d.description ?? '');
-                        setTaxable(d.taxable);
-                        const filled = [
-                          t('quotes.editor.autoFill.filledName'),
-                          t('quotes.editor.autoFill.filledDescription'),
-                          d.taxable ? t('quotes.editor.autoFill.filledTaxableOn') : t('quotes.editor.autoFill.filledTaxableOff'),
-                        ];
-                        // Pre-fill cost/price only into untouched fields — auto-fill
-                        // must never overwrite a number the user already typed (read
-                        // the refs: the lookup takes seconds and state may have moved).
-                        if (result.estimatedCost != null && costRef.current.trim() === '') {
-                          const c = result.estimatedCost.toFixed(2);
-                          setCost(c);
-                          filled.push(t('quotes.editor.autoFill.filledEstimatedCost', { amount: formatMoney(Number(c), currency) }));
-                          if (defaultMarkupPct != null && (priceRef.current.trim() === '' || Number(priceRef.current) === 0)) {
-                            const p = priceFromMarkup(c, defaultMarkupPct);
-                            setPrice(p);
-                            setMarkup(String(Number(defaultMarkupPct)));
-                            priceAuthority.current = 'markup';
-                            filled.push(t('quotes.editor.autoFill.filledPriceWithDefaultMarkup', { amount: formatMoney(Number(p), currency), markup: String(Number(defaultMarkupPct)) }));
-                          }
-                        }
-                        setAutoFilled(filled);
-                      }}
-                    />
-                    {(name.trim() || desc.trim()) && (
-                      <PolishButton
-                        idSuffix={`quote-manual-${block.id}`}
-                        getText={() => ({ name, description: desc })}
-                        onApply={(r) => {
-                          if (r.name !== null) setName(r.name);
-                          if (r.description !== null) setDesc(r.description);
-                        }}
-                      />
-                    )}
-                  </div>
-                  {/* What the last auto-fill actually touched — the AI applies
-                      directly to the form, so the user must be told what changed. */}
-                  {autoFilled && (
-                    <p role="status" data-testid={`quote-manual-autofilled-${block.id}`} className="text-xs text-muted-foreground">
-                      <span className="font-medium text-foreground">{t('quotes.editor.autoFill.label')}</span> {autoFilled.join(' · ')}. {t('quotes.editor.autoFill.review')}
-                    </p>
-                  )}
-                  <input
-                    type="text" placeholder={t('quotes.editor.line.namePlaceholder')} aria-label={t('quotes.editor.line.nameAria')} value={name}
-                    onChange={(e) => setName(e.target.value)}
-                    data-testid={`quote-manual-name-${block.id}`}
-                    className="h-9 w-full rounded-md border bg-background px-3 text-sm focus:outline-hidden focus:ring-2 focus:ring-ring"
-                  />
-                  <div className="grid grid-cols-1 items-start gap-2 sm:grid-cols-[1fr_70px_90px_110px]">
-                    <label className="block">
-                      <span className="mb-1 block text-xs text-muted-foreground">{t('quotes.editor.line.descriptionOptional')}</span>
-                      <textarea
-                        value={desc}
-                        onChange={(e) => setDesc(e.target.value)}
-                        rows={2}
-                        data-testid={`quote-manual-desc-${block.id}`}
-                        className="min-h-9 w-full resize-y rounded-md border bg-background px-3 py-2 text-sm focus:outline-hidden focus:ring-2 focus:ring-ring"
-                      />
-                    </label>
-                    <label className="block">
-                      <span className="mb-1 block text-xs text-muted-foreground">{t('quotes.editor.table.qty')}</span>
-                      <input
-                        type="number" min="1" step="1" value={qty}
-                        onChange={(e) => { setQty(e.target.value); clearManualError('qty'); }}
-                        aria-invalid={manualErrors.qty ? true : undefined}
-                        aria-describedby={manualErrors.qty ? `quote-manual-qty-error-${block.id}` : undefined}
-                        data-testid={`quote-manual-qty-${block.id}`}
-                        className={`h-9 w-full rounded-md border bg-background px-3 text-sm focus:outline-hidden focus:ring-2 focus:ring-ring ${manualErrors.qty ? 'border-destructive' : ''}`}
-                      />
-                      {manualErrors.qty && (
-                        <span id={`quote-manual-qty-error-${block.id}`} className="mt-0.5 block text-xs text-destructive" data-testid={`quote-manual-qty-error-${block.id}`}>
-                          {manualErrors.qty}
-                        </span>
-                      )}
-                    </label>
-                    <label className="block">
-                      <span className="mb-1 block text-xs text-muted-foreground">{t('quotes.editor.table.unitPrice')}</span>
-                      <input
-                        type="number" min="0" step="0.01" value={price}
-                        onChange={(e) => { onPriceChange(e.target.value); clearManualError('price'); }}
-                        aria-invalid={manualErrors.price ? true : undefined}
-                        aria-describedby={manualErrors.price ? `quote-manual-price-error-${block.id}` : undefined}
-                        data-testid={`quote-manual-price-${block.id}`}
-                        className={`h-9 w-full rounded-md border bg-background px-3 text-sm focus:outline-hidden focus:ring-2 focus:ring-ring ${manualErrors.price ? 'border-destructive' : ''}`}
-                      />
-                      {manualErrors.price && (
-                        <span id={`quote-manual-price-error-${block.id}`} className="mt-0.5 block text-xs text-destructive" data-testid={`quote-manual-price-error-${block.id}`}>
-                          {manualErrors.price}
-                        </span>
-                      )}
-                    </label>
-                    <label className="block">
-                      <span className="mb-1 block text-xs text-muted-foreground">{t('quotes.editor.line.billing')}</span>
-                      <select
-                        value={recurrence}
-                        onChange={(e) => setRecurrence(e.target.value as QuoteLineRecurrence)}
-                        data-testid={`quote-manual-recurrence-${block.id}`}
-                        className="h-9 w-full rounded-md border bg-background px-2 text-sm focus:outline-hidden focus:ring-2 focus:ring-ring"
-                      >
-                        <option value="one_time">{t('quotes.editor.recurrence.one_time')}</option>
-                        <option value="monthly">{t('quotes.editor.recurrence.monthly')}</option>
-                        <option value="annual">{t('quotes.editor.recurrence.annual')}</option>
-                      </select>
-                    </label>
-                  </div>
-                  {/* Internal-only cost & identity fields (never shown to the customer).
-                      Divider + top padding sets them apart from the customer-facing
-                      fields above so the two groups don't read as one dense block. */}
-                  <p className="mt-1 border-t pt-3 text-xs font-medium uppercase tracking-wide text-muted-foreground">{t('quotes.editor.internal.full')}</p>
-                  <div className="grid grid-cols-1 gap-2 sm:grid-cols-[1fr_1fr_110px_100px]">
-                    <label className="block">
-                      <span className="mb-1 block text-xs text-muted-foreground">{t('quotes.editor.line.skuOptional')}</span>
-                      <input
-                        type="text" value={sku}
-                        onChange={(e) => setSku(e.target.value)}
-                        data-testid={`quote-manual-sku-${block.id}`}
-                        className="h-9 w-full rounded-md border bg-background px-3 text-sm focus:outline-hidden focus:ring-2 focus:ring-ring"
-                      />
-                    </label>
-                    <label className="block">
-                      <span className="mb-1 block text-xs text-muted-foreground">{t('quotes.editor.line.partNumberOptional')}</span>
-                      <input
-                        type="text" value={partNumber}
-                        onChange={(e) => setPartNumber(e.target.value)}
-                        data-testid={`quote-manual-partnumber-${block.id}`}
-                        className="h-9 w-full rounded-md border bg-background px-3 text-sm focus:outline-hidden focus:ring-2 focus:ring-ring"
-                      />
-                    </label>
-                    <label className="block">
-                      <span className="mb-1 block text-xs text-muted-foreground">{t('quotes.editor.line.unitCost')}</span>
-                      <input
-                        type="number" min="0" step="0.01" value={cost}
-                        onChange={(e) => { onCostChange(e.target.value); clearManualError('cost'); }}
-                        aria-invalid={manualErrors.cost ? true : undefined}
-                        aria-describedby={manualErrors.cost ? `quote-manual-cost-error-${block.id}` : undefined}
-                        data-testid={`quote-manual-cost-${block.id}`}
-                        className={`h-9 w-full rounded-md border bg-background px-3 text-right text-sm tabular-nums focus:outline-hidden focus:ring-2 focus:ring-ring ${manualErrors.cost ? 'border-destructive' : ''}`}
-                      />
-                      {manualErrors.cost && (
-                        <span id={`quote-manual-cost-error-${block.id}`} className="mt-0.5 block text-xs text-destructive" data-testid={`quote-manual-cost-error-${block.id}`}>
-                          {manualErrors.cost}
-                        </span>
-                      )}
-                    </label>
-                    <label className="block">
-                      <span className="mb-1 block text-xs text-muted-foreground">{t('quotes.editor.line.markupPercent')}</span>
-                      <input
-                        type="number" step="0.1" value={markup}
-                        onChange={(e) => onMarkupChange(e.target.value)}
-                        disabled={cost.trim() === ''}
-                        // Sighted users can see the empty cost field; AT users can't —
-                        // mirror the edit-row band's disabled-reason wiring.
-                        title={cost.trim() === '' ? t('quotes.editor.line.enterCostFirstMarkup') : undefined}
-                        aria-describedby={cost.trim() === '' ? `quote-manual-markup-hint-${block.id}` : undefined}
-                        data-testid={`quote-manual-markup-${block.id}`}
-                        className="h-9 w-full rounded-md border bg-background px-3 text-right text-sm tabular-nums focus:outline-hidden focus:ring-2 focus:ring-ring disabled:opacity-60"
-                      />
-                      {cost.trim() === '' && <span id={`quote-manual-markup-hint-${block.id}`} className="sr-only">{t('quotes.editor.line.enterCostFirstMarkupSentence')}</span>}
-                    </label>
-                  </div>
-                  <div className="flex flex-wrap items-center justify-between gap-2">
-                    <div className="flex flex-wrap items-center gap-3 text-xs">
-                      <label className="flex items-center gap-1">
-                        <input type="checkbox" checked={taxable} onChange={(e) => setTaxable(e.target.checked)} data-testid={`quote-manual-taxable-${block.id}`} />
-                        {t('quotes.editor.table.taxable')}
-                      </label>
-                      <label className="flex items-center gap-1">
-                        <input type="checkbox" checked={saveToCatalog} onChange={(e) => setSaveToCatalog(e.target.checked)} data-testid={`quote-manual-save-catalog-${block.id}`} />
-                        {t('quotes.editor.line.saveToCatalog')}
-                      </label>
-                    </div>
-                    <button
-                      type="button"
-                      onClick={() => void submitManual()}
-                      // A line needs a name OR a description (mirrors the API + addManual
-                      // refine). Gating on description alone silently blocked valid
-                      // name-only lines like a titled SKU with no prose.
-                      disabled={addLineBusy || (!name.trim() && !desc.trim())}
-                      aria-busy={addLineBusy}
-                      data-testid={`quote-manual-add-${block.id}`}
-                      className="inline-flex h-9 items-center justify-center gap-1.5 rounded-md bg-primary px-4 text-sm font-medium text-primary-foreground hover:opacity-90 disabled:opacity-50"
-                    >
-                      {addLineBusy && <Loader2 className="h-3.5 w-3.5 animate-spin" aria-hidden="true" />}
-                      {addLineBusy ? t('quotes.editor.actions.adding') : t('quotes.editor.actions.addLine')}
-                    </button>
-                  </div>
-                </div>
-              )}
-            </div>
-            )}
-          </div>
-        )}
-      </div>
-    </section>
-  );
-}
-
-// ── A single read-only pricing-table line (no write permission) ───────────
-// Mirrors EditableLineRow's two-row shape — the customer-facing cells plus the
-// internal cost/markup/net band — but renders everything as plain text.
-function ReadonlyLineRow({ line: l, quoteId, currency, taxRate, isFirst, showInternal }: { line: QuoteLine; quoteId: string; currency: string; taxRate: string | null; isFirst: boolean; showInternal: boolean }) {
-  const { t } = useTranslation('billing');
-  const mk = markupPct(l.unitPrice, l.unitCost);
-  const markupStr = mk === null ? t('quotes.editor.symbols.notAvailable') : formatPercent(mk / 100, { maximumFractionDigits: 2 });
-  const netCents = l.unitCost === null
-    ? null
-    : toCents(computeLineTotal(l.quantity, l.unitPrice)) - toCents(computeLineTotal(l.quantity, l.unitCost));
-  const tax = lineTaxAmount(l.lineTotal, l.taxable, taxRate);
-  // Billing cadence rides in the money cells ('/mo', '/yr'); one-time is unmarked.
-  const suffix = l.recurrence === 'monthly' ? t('quotes.editor.units.perMonth') : l.recurrence === 'annual' ? t('quotes.editor.units.perYear') : '';
-  return (
-    <>
-      <tr className="border-t [&>td]:pt-4" data-testid={`quote-line-${l.id}`}>
-        <td className="px-1.5 py-2">
-          <div className="flex items-start gap-2">
-            {l.imageId
-              ? <LineImageThumb quoteId={quoteId} imageId={l.imageId} />
-              : l.catalogItemId && <CatalogLineThumb catalogItemId={l.catalogItemId} />}
-            <div>
-              <div className="font-medium">{lineTitle(l)}</div>
-              {lineBlurb(l) && <div className="whitespace-pre-line text-xs text-muted-foreground">{lineBlurb(l)}</div>}
-            </div>
-          </div>
-        </td>
-        <td className="px-1.5 py-2 text-right tabular-nums">{formatQuantity(l.quantity)}</td>
-        <td className="whitespace-nowrap px-1.5 py-2 text-right tabular-nums">
-          {formatMoney(l.unitPrice, currency)}{suffix && <span className="text-xs text-muted-foreground">{suffix}</span>}
-        </td>
-        <td className="whitespace-nowrap px-1.5 py-2 text-right tabular-nums">
-          {formatMoney(l.lineTotal, currency)}{suffix && <span className="text-xs text-muted-foreground">{suffix}</span>}
-          <div className="text-xs font-normal text-muted-foreground" data-testid={`quote-line-tax-${l.id}`}>
-            {tax !== null
-              ? t('quotes.editor.table.taxSuffix', { amount: formatMoney(tax, currency) })
-              : l.taxable ? t('quotes.editor.table.taxable') : null}
-          </div>
-        </td>
-      </tr>
-      <tr className={`border-0 ${showInternal ? '' : 'hidden'}`} data-testid={`quote-line-internal-${l.id}`}>
-        <td colSpan={4} className="px-2 pb-2">
-          <div className="flex flex-wrap items-center gap-x-3 gap-y-1 rounded-md bg-muted/40 px-2 py-1 text-xs text-[hsl(220_12%_40%)] dark:text-muted-foreground">
-            {/* Full disclaimer on the first row, a subtle "Internal" tag on the rest. */}
-            <span className="font-medium uppercase tracking-wide">{isFirst ? t('quotes.editor.internal.full') : t('quotes.editor.internal.short')}</span>
-            <span data-testid={`quote-line-sku-${l.id}`}>{t('quotes.editor.line.sku')} {l.sku || t('quotes.editor.symbols.notAvailable')}</span>
-            <span data-testid={`quote-line-partnumber-${l.id}`}>{t('quotes.editor.line.partNumberAbbr')} {l.partNumber || t('quotes.editor.symbols.notAvailable')}</span>
-            <span data-testid={`quote-line-cost-${l.id}`}>{t('quotes.editor.line.cost')} {l.unitCost === null ? t('quotes.editor.symbols.notAvailable') : formatMoney(l.unitCost, currency)}</span>
-            <span data-testid={`quote-line-markup-${l.id}`}>{t('quotes.editor.line.markup')} {markupStr}</span>
-            <span className="ml-auto">{t('quotes.editor.line.profit')}{' '}
-              <span className="font-medium tabular-nums text-foreground" data-testid={`quote-line-net-${l.id}`}>
-                {netCents === null ? t('quotes.editor.symbols.notAvailable') : formatMoney(fromCents(netCents), currency)}
-              </span>
-            </span>
-          </div>
-        </td>
-      </tr>
-    </>
-  );
-}
-
-// ── A single editable pricing-table line (writers only) ───────────────────
-// Each field is locally controlled and committed on blur (text/number) or on
-// change (taxable checkbox, recurrence select) — but only when the value
-// actually differs from the persisted line, so a focus-without-edit doesn't
-// fire a redundant PATCH. The parent's onEdit routes through updateLine +
-// runAction and then refresh()es, which re-pulls the line; we resync local
-// state to the incoming prop so server-side normalization (e.g. recomputed
-// totals, clamped quantity) wins.
-function EditableLineRow({
-  line, quoteId, currency, taxRate, isPending, isFirst, isLast, showInternal, depositSelectMode, onEdit, onMove, onRemove, onDraft,
-  moveTargets, onMoveTo,
-}: {
-  line: QuoteLine;
-  quoteId: string;
-  currency: string;
-  taxRate: string | null;
-  isPending: (key: string) => boolean;
-  isFirst: boolean;
-  isLast: boolean;
-  showInternal: boolean;
-  /** Show the deposit-eligible checkbox (quote deposit = 'selected_lines'). */
-  depositSelectMode: boolean;
-  onEdit: (lineId: string, body: LineUpdate, scopeKey?: string) => Promise<boolean>;
-  onMove: (line: QuoteLine, direction: 'up' | 'down') => void;
-  onRemove: (line: QuoteLine) => void;
-  onDraft: (lineId: string, draft: QuoteLineForMath | null) => void;
-  /** Other pricing panels (empty → the Move-to control is hidden). */
-  moveTargets: { id: string; label: string }[];
-  onMoveTo: (line: QuoteLine, targetBlockId: string) => void;
-}) {
-  const { t } = useTranslation('billing');
-  // Per-field pending: only the in-flight control disables, so a slow qty save
-  // never freezes price/name/desc (the scoped-pending backport — InvoiceEditor's
-  // LineRow got this first). Remove keeps the whole-row key: the confirm-dialog
-  // removal flow runs under `line:<id>` and should hold the row's actions.
-  const fieldBusy = (field: string) => isPending(`line:${line.id}:${field}`);
-  const removeBusy = isPending(`line:${line.id}`);
-  const [name, setName] = useState(line.name ?? '');
-  const [desc, setDesc] = useState(line.description ?? '');
-  // Rest-state density: an EMPTY description renders no textarea — a compact
-  // "+ description" affordance opens it. Lines with prose keep it open (and a
-  // server resync that lands prose re-opens it).
-  const [descOpen, setDescOpen] = useState(Boolean((line.description ?? '').trim()));
-  useEffect(() => { if ((line.description ?? '').trim()) setDescOpen(true); }, [line.description]);
-  // Quantity edits/displays in its bare form ('3', not the stored '3.00') so the
-  // input matches its own step="1" and the customer-facing rendering.
-  const [qty, setQty] = useState(formatQuantity(line.quantity));
-  const [price, setPrice] = useState(line.unitPrice);
-  // recurrence/taxable are committed on change (not blur); keep them in local
-  // state so the control updates instantly rather than lagging until the
-  // refresh() round-trip lands, and revert if the save fails.
-  const [rec, setRec] = useState(line.recurrence);
-  const [taxable, setTaxable] = useState(line.taxable);
-  // Deposit-eligibility is committed on change (like taxable) and reverts on a
-  // failed save; resynced from the server prop after each refresh().
-  const [depositEligible, setDepositEligible] = useState(line.depositEligible ?? false);
-  // Internal cost/identity fields (cost drives the markup/net strip below the row).
-  const [cost, setCost] = useState(line.unitCost ?? '');
-  const [sku, setSku] = useState(line.sku ?? '');
-  const [partNumber, setPartNumber] = useState(line.partNumber ?? '');
-
-  // "Move to…" menu. Fixed-position so the overflow-x-auto table wrapper can't
-  // clip it; closes on outside click or Escape (which refocuses the trigger —
-  // focus moved into the menu on open, so it would otherwise drop to <body>).
-  const [movePos, setMovePos] = useState<{ top: number; left: number } | null>(null);
-  const moveMenuRef = useRef<HTMLDivElement>(null);
-  const moveTriggerRef = useRef<HTMLButtonElement>(null);
-  const { listRef: moveListRef, onKeyDown: onMoveListKeyDown } = useMenuKeyboard(movePos !== null, () => setMovePos(null));
-  useEffect(() => {
-    if (!movePos) return;
-    const onDown = (e: MouseEvent) => {
-      if (moveMenuRef.current && !moveMenuRef.current.contains(e.target as Node)) setMovePos(null);
-    };
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') { setMovePos(null); moveTriggerRef.current?.focus(); }
-    };
-    document.addEventListener('mousedown', onDown);
-    document.addEventListener('keydown', onKey);
-    return () => { document.removeEventListener('mousedown', onDown); document.removeEventListener('keydown', onKey); };
-  }, [movePos]);
-
-  // Resync the typed fields from the server, but never over an edit in progress.
-  // We track "has the user typed since the last commit?" rather than comparing
-  // values: local state holds a raw string ('9.999') while the prop is a
-  // formatted decimal ('10.00'), so a value comparison both (a) fails to re-adopt
-  // a server-normalized value — leaving the row stuck amber-dirty showing a wrong
-  // optimistic total when the server rounds — and (b) is fragile across formats.
-  // The flag is set on keystroke and cleared when a commit is initiated (on blur), so:
-  //   • a quiet/leading-edge refresh landing mid-type keeps the user's keystrokes
-  //     ("edit qty→5, blur, type 7" never loses the 7), and
-  //   • after the user stops editing, the next prop adopts the server's canonical
-  //     value (e.g. 9.999 → 10.00), clearing the dirty ring and the optimism.
-  const nameEdited = useRef(false);
-  const descEdited = useRef(false);
-  // Auto-grow the (full-width) description textarea to fit its content, while
-  // still allowing the user to drag the resize handle for a bigger/smaller box.
-  const descRef = useRef<HTMLTextAreaElement>(null);
-  const autoGrowDesc = () => {
-    const el = descRef.current;
-    if (!el) return;
-    el.style.height = 'auto';
-    el.style.height = `${el.scrollHeight}px`;
-  };
-  const qtyEdited = useRef(false);
-  const priceEdited = useRef(false);
-  const costEdited = useRef(false);
-  const skuEdited = useRef(false);
-  const partEdited = useRef(false);
-  useEffect(() => { if (!nameEdited.current) setName(line.name ?? ''); }, [line.name]);
-  useEffect(() => { if (!descEdited.current) setDesc(line.description ?? ''); }, [line.description]);
-  // Re-fit the description box after any value change (typing or server resync).
-  useEffect(() => { autoGrowDesc(); }, [desc]);
-  useEffect(() => { if (!qtyEdited.current) setQty(formatQuantity(line.quantity)); }, [line.quantity]);
-  useEffect(() => { if (!priceEdited.current) setPrice(line.unitPrice); }, [line.unitPrice]);
-  useEffect(() => { if (!costEdited.current) setCost(line.unitCost ?? ''); }, [line.unitCost]);
-  useEffect(() => { if (!skuEdited.current) setSku(line.sku ?? ''); }, [line.sku]);
-  useEffect(() => { if (!partEdited.current) setPartNumber(line.partNumber ?? ''); }, [line.partNumber]);
-  // recurrence/taxable are committed on change (the PATCH resolves before the
-  // refresh GET fires), so a stale resync can't race them — a plain resync wins.
-  useEffect(() => { setRec(line.recurrence); }, [line.recurrence]);
-  useEffect(() => { setTaxable(line.taxable); }, [line.taxable]);
-  useEffect(() => { setDepositEligible(line.depositEligible ?? false); }, [line.depositEligible]);
-
-  // Inline validation errors for the money/qty fields. A rejected entry keeps
-  // the user's input in the field (to correct, not re-type) with the message
-  // rendered directly under the row — replacing the old toast-1000px-away +
-  // silent snap-back. Same visual contract as the contract-variable errors
-  // (aria-invalid + destructive border + text). Cleared on the next keystroke.
-  // Server-side rejections still surface through runAction's toast.
-  const [fieldErrors, setFieldErrors] = useState<{ qty?: string; price?: string; cost?: string }>({});
-  const setFieldError = useCallback((field: 'qty' | 'price' | 'cost', msg: string | null) => {
-    setFieldErrors((e) => {
-      if (msg === null) {
-        if (!(field in e)) return e;
-        const n = { ...e }; delete n[field]; return n;
-      }
-      return { ...e, [field]: msg };
-    });
-  }, []);
-
-  // Quiet "Saved" flash in place of the old per-field success toast. This is a
-  // single row-level flag on purpose: committing any one field briefly pulses the
-  // green ring across the row's fields, reading as "this line saved" rather than
-  // tracking which individual cell changed.
-  const [saved, setSaved] = useState(false);
-  const savedTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
-  useEffect(() => () => { if (savedTimer.current) clearTimeout(savedTimer.current); }, []);
-  const flashSaved = useCallback(() => {
-    setSaved(true);
-    if (savedTimer.current) clearTimeout(savedTimer.current);
-    savedTimer.current = setTimeout(() => setSaved(false), 1500);
-  }, []);
-  // `field` scopes the pending key to the one control being committed; commits
-  // without a field (none today) would fall back to freezing the whole row.
-  const edit = useCallback(async (body: LineUpdate, field?: string): Promise<boolean> => {
-    const ok = await onEdit(line.id, body, field ? `line:${line.id}:${field}` : undefined);
-    if (ok) flashSaved();
-    return ok;
-  }, [onEdit, line.id, flashSaved]);
-
-  // Per-line product image: resolve an imageId from an uploaded file OR a pasted
-  // URL (the server copies the bytes in — not a hotlink), then PATCH the line's
-  // imageId. Local busy (not the pending map) because the upload itself runs
-  // before any line mutation exists to scope. The URL path is a disclosure — a
-  // "From URL" button reveals `imageUrlOpen`'s inline field — rather than a
-  // per-row tab toggle, which would bloat every line; the add-block form (which
-  // has room) uses tabs instead.
-  const imageInputRef = useRef<HTMLInputElement>(null);
-  const [imageBusy, setImageBusy] = useState(false);
-  const [imageUrlOpen, setImageUrlOpen] = useState(false);
-  const [imageUrlDraft, setImageUrlDraft] = useState('');
-  // Attach `uploaded.imageId` to the line, and only on a successful PATCH reset the
-  // URL disclosure. Shared by both the file and URL paths so success behaves
-  // identically. `edit` swallows a failed PATCH inside runScoped (toasts, returns
-  // false, never throws), so gating the reset on its result keeps the disclosure
-  // open with the URL intact on failure — otherwise the panel would collapse and
-  // wipe the URL exactly as if it had saved, contradicting the error toast.
-  const applyImageId = useCallback(async (imageId: string) => {
-    if (await edit({ imageId }, 'image')) {
-      setImageUrlDraft('');
-      setImageUrlOpen(false);
-    }
-  }, [edit]);
-  const attachImage = useCallback((file: File) => {
-    if (file.size > 5 * 1024 * 1024) {
-      handleActionError(new Error('image too large'), t('quotes.editor.errors.imageTooLarge'));
-      return;
-    }
-    void (async () => {
-      setImageBusy(true);
-      try {
-        const uploaded = await runAction<{ imageId: string }>({
-          request: () => uploadQuoteImage(quoteId, file),
-          errorFallback: t('quotes.editor.errors.uploadImage'),
-          onUnauthorized: UNAUTHORIZED,
-          parseSuccess: (d) => (d as { data: { imageId: string } }).data,
-        });
-        await applyImageId(uploaded.imageId);
-      } catch {
-        // runAction already surfaced the failure (toast/redirect).
-      } finally {
-        setImageBusy(false);
-      }
-    })();
-  }, [quoteId, applyImageId, t]);
-  // URL path: the server fetches + copies the remote bytes (SSRF-guarded, 5 MB
-  // cap enforced server-side — unlike the file path there's no local size check
-  // because the bytes aren't in hand here). Mirrors the image block's URL flow.
-  const attachImageFromUrl = useCallback(() => {
-    const url = imageUrlDraft.trim();
-    if (!url) return;
-    void (async () => {
-      setImageBusy(true);
-      try {
-        const uploaded = await runAction<{ imageId: string }>({
-          request: () => addQuoteImageFromUrl(quoteId, url),
-          errorFallback: t('quotes.editor.errors.fetchImageFromUrl'),
-          onUnauthorized: UNAUTHORIZED,
-          parseSuccess: (d) => (d as { data: { imageId: string } }).data,
-        });
-        await applyImageId(uploaded.imageId);
-      } catch {
-        // runAction already surfaced the failure (toast/redirect).
-      } finally {
-        setImageBusy(false);
-      }
-    })();
-  }, [quoteId, imageUrlDraft, applyImageId, t]);
-  // Per-field dirty cue (mirrors the terms/tax ring) so every editable surface
-  // signals unsaved state the same way.
-  const nameDirty = name.trim() !== (line.name ?? '');
-  const descDirty = desc.trim() !== (line.description ?? '');
-  const qtyDirty = Number(qty) !== Number(line.quantity);
-  const priceDirty = Number(price) !== Number(line.unitPrice);
-
-  // Effective qty/price for the optimistic Total/Tax and the rail draft. A blank
-  // or non-positive qty / negative price isn't an optimism input — it would flash
-  // the line to $0 while the user is mid-retype (and `commitQty` rejects qty ≤ 0
-  // anyway), so fall back to the persisted value until the field holds a real one.
-  const qtyNum = Number(qty);
-  const priceNum = Number(price);
-  const qtyValid = qty.trim() !== '' && Number.isFinite(qtyNum) && qtyNum > 0;
-  const priceValid = price.trim() !== '' && Number.isFinite(priceNum) && priceNum >= 0;
-  const effQty = qtyValid ? qty : line.quantity;
-  const effPrice = priceValid ? price : line.unitPrice;
-  const totalDiverged = Number(effQty) !== Number(line.quantity) || Number(effPrice) !== Number(line.unitPrice);
-
-  // The row's Total/Tax use the SAME shared cents math as the rail
-  // (computeLineTotal — round-half-up at the cent boundary), so a sub-cent unit
-  // price can't make the row Total and the rail contribution disagree by a cent
-  // while typing. When qty/price are unchanged we defer to the authoritative
-  // persisted lineTotal so server normalization still wins on settle.
-  const displayTotal = totalDiverged ? computeLineTotal(effQty, effPrice) : line.lineTotal;
-  const displayTax = lineTaxAmount(displayTotal, taxable, taxRate);
-
-  // Markup is derived from price+cost. The input is controlled by local state that
-  // resyncs from the derived value when price/cost change — but only while the
-  // field is NOT focused, so a cross-field cost edit never yanks the caret. (This
-  // replaces the old key={markupStr} remount, which dropped focus on every
-  // commit.) Net is (price − cost) × qty in cents; "—" when no cost is set.
-  const mk = markupPct(effPrice, cost);
-  const markupStr = mk === null ? '' : String(Number(mk.toFixed(2)));
-  const markupFocused = useRef(false);
-  const [markupInput, setMarkupInput] = useState(markupStr);
-  useEffect(() => { if (!markupFocused.current) setMarkupInput(markupStr); }, [markupStr]);
-  const netCents = cost.trim() === ''
-    ? null
-    : toCents(computeLineTotal(effQty, effPrice)) - toCents(computeLineTotal(effQty, cost));
-  const costDirty = cost.trim() === '' ? line.unitCost !== null : Number(cost) !== Number(line.unitCost);
-  const skuDirty = sku.trim() !== (line.sku ?? '');
-  const partDirty = partNumber.trim() !== (line.partNumber ?? '');
-
-  // Report this row's effective values to the parent so the rail "Live totals"
-  // recompute uses the same inputs. Emit null once nothing diverges, so the rail
-  // reverts to the authoritative server figures; cleanup on unmount avoids a
-  // phantom draft skewing the rail after a delete.
-  const depositEligibleDirty = depositEligible !== (line.depositEligible ?? false);
-  const diverged = totalDiverged || taxable !== line.taxable || rec !== line.recurrence || costDirty || depositEligibleDirty;
-  useEffect(() => {
-    onDraft(line.id, diverged
-      ? { quantity: String(effQty), unitPrice: String(effPrice), unitCost: cost || null, taxable, customerVisible: line.customerVisible, recurrence: rec, depositEligible, itemType: line.itemType ?? null }
-      : null);
-  }, [onDraft, line.id, line.customerVisible, line.itemType, diverged, effQty, effPrice, cost, taxable, rec, depositEligible]);
-  // Clear this row's draft when it unmounts (e.g. removed) so the rail doesn't
-  // keep a phantom override.
-  useEffect(() => () => onDraft(line.id, null), [onDraft, line.id]);
-
-  const commitName = () => {
-    const next = name.trim();
-    nameEdited.current = false; // committing — let the server value re-adopt next
-    if (next === (line.name ?? '')) { setName(line.name ?? ''); return; }
-    // A line can't have both name and description blank (mirrors the API refine).
-    if (!next && !(line.description ?? '').trim()) {
-      handleActionError(new Error('empty line'), t('quotes.editor.errors.lineNeedsNameOrDescription'));
-      setName(line.name ?? '');
-      return;
-    }
-    void edit({ name: next || null }, 'name');
-  };
-  const commitDesc = () => {
-    const next = desc.trim();
-    descEdited.current = false; // committing — let the server value re-adopt next
-    if (next === (line.description ?? '')) { setDesc(line.description ?? ''); return; }
-    if (!next && !(line.name ?? '').trim()) {
-      handleActionError(new Error('empty line'), t('quotes.editor.errors.lineNeedsNameOrDescription'));
-      setDesc(line.description ?? '');
-      return;
-    }
-    void edit({ description: next || null }, 'desc');
-  };
-  const commitQty = () => {
-    const n = Number(qty);
-    qtyEdited.current = false;
-    if (n === Number(line.quantity)) { setQty(formatQuantity(line.quantity)); setFieldError('qty', null); return; } // unchanged — silent
-    // A rejected entry stays in the field with an inline error under the row —
-    // never a far-away toast plus a silent snap-back. The optimistic Total/rail
-    // already fall back to the persisted value while the field holds an invalid
-    // one (qtyValid), so nothing downstream computes from the bad input.
-    if (!Number.isFinite(n) || n <= 0 || !Number.isInteger(n)) {
-      setFieldError('qty', t('quotes.editor.errors.quantityWholeGreaterThanZero'));
-      return;
-    }
-    setFieldError('qty', null);
-    void edit({ quantity: n }, 'qty');
-  };
-  const commitPrice = () => {
-    const n = Number(price);
-    priceEdited.current = false;
-    if (n === Number(line.unitPrice)) { setPrice(line.unitPrice); setFieldError('price', null); return; } // unchanged — silent
-    if (!Number.isFinite(n) || n < 0) {
-      setFieldError('price', t('quotes.editor.errors.unitPriceZeroOrMore'));
-      return;
-    }
-    setFieldError('price', null);
-    void edit({ unitPrice: n }, 'price');
-  };
-  const commitCost = () => {
-    costEdited.current = false;
-    if (cost.trim() === '') { setFieldError('cost', null); if (line.unitCost !== null) void edit({ unitCost: null }, 'cost'); return; }
-    const n = Number(cost);
-    if (!Number.isFinite(n) || n < 0) {
-      setFieldError('cost', t('quotes.editor.errors.costZeroOrMore'));
-      return;
-    }
-    setFieldError('cost', null);
-    if (n !== Number(line.unitCost)) void edit({ unitCost: n }, 'cost');
-  };
-  const commitSku = () => {
-    skuEdited.current = false;
-    const next = sku.trim();
-    if (next !== (line.sku ?? '')) void edit({ sku: next || null }, 'sku');
-  };
-  const commitPartNumber = () => {
-    partEdited.current = false;
-    const next = partNumber.trim();
-    if (next !== (line.partNumber ?? '')) void edit({ partNumber: next || null }, 'pn');
-  };
-  // Editing markup% commits a new unit price derived from cost: price = cost·(1+m).
-  const onMarkupCommit = (raw: string) => {
-    const m = Number(raw);
-    // Need a cost base, and treat an emptied markup field as "leave price alone" —
-    // Number('') is 0 (finite), which would otherwise rewrite unitPrice down to cost
-    // (zero margin) just because the user cleared the field.
-    if (cost.trim() === '' || raw.trim() === '' || !Number.isFinite(m)) return;
-    const nextPrice = priceFromMarkup(cost, m);
-    setPrice(nextPrice);
-    priceEdited.current = false;
-    if (Number(nextPrice) !== Number(line.unitPrice)) void edit({ unitPrice: Number(nextPrice) }, 'price');
-  };
-
-  return (
-    <>
-    <tr className="border-t align-top [&>td]:pt-4" data-testid={`quote-line-${line.id}`}>
-      {/* Column min-width (min-w-[12rem]) is declared on the cell so table
-          auto-layout reserves the name column instead of squeezing it below the
-          input's width (which used to overflow into the qty cell). */}
-      <td className="min-w-[12rem] px-1.5 py-2">
-        <div className="flex min-w-0 items-start gap-2">
-          {line.imageId
-            ? <LineImageThumb quoteId={quoteId} imageId={line.imageId} />
-            : line.catalogItemId && <CatalogLineThumb catalogItemId={line.catalogItemId} />}
-          <div className="w-full min-w-0 space-y-1">
-            <input
-              type="text"
-              value={name}
-              aria-label={t('quotes.editor.line.nameAria')}
-              placeholder={t('quotes.editor.line.namePlaceholder')}
-              onChange={(e) => { setName(e.target.value); nameEdited.current = true; }}
-              onBlur={commitName}
-              disabled={fieldBusy('name')}
-              data-testid={`quote-line-name-${line.id}`}
-              className={`h-9 w-full rounded-md border bg-transparent px-2 py-1 text-sm font-medium transition-colors focus:outline-hidden disabled:opacity-60 ${seamless(fieldRing(nameDirty, saved))}`}
-            />
-          </div>
-        </div>
-      </td>
-      <td className="px-1.5 py-2 text-right">
-        <input
-          type="number" min="1" step="1"
-          value={qty}
-          aria-label={t('quotes.editor.line.quantityAria')}
-          onChange={(e) => { setQty(e.target.value); qtyEdited.current = true; setFieldError('qty', null); }}
-          onBlur={commitQty}
-          disabled={fieldBusy('qty')}
-          aria-invalid={fieldErrors.qty ? true : undefined}
-          aria-describedby={fieldErrors.qty ? `quote-line-qty-error-${line.id}` : undefined}
-          data-testid={`quote-line-qty-${line.id}`}
-          className={`h-9 w-14 rounded-md border bg-transparent px-2 text-right text-sm tabular-nums transition-colors focus:outline-hidden disabled:opacity-60 ${fieldErrors.qty ? 'border-destructive' : seamless(fieldRing(qtyDirty, saved))}`}
-        />
-      </td>
-      <td className="px-1.5 py-2 text-right">
-        <input
-          type="number" min="0" step="0.01"
-          value={price}
-          aria-label={t('quotes.editor.table.unitPrice')}
-          onChange={(e) => { setPrice(e.target.value); priceEdited.current = true; setFieldError('price', null); }}
-          onBlur={commitPrice}
-          disabled={fieldBusy('price')}
-          aria-invalid={fieldErrors.price ? true : undefined}
-          aria-describedby={fieldErrors.price ? `quote-line-price-error-${line.id}` : undefined}
-          data-testid={`quote-line-price-${line.id}`}
-          className={`h-9 w-24 rounded-md border bg-transparent px-2 text-right text-sm tabular-nums transition-colors focus:outline-hidden disabled:opacity-60 ${fieldErrors.price ? 'border-destructive' : seamless(fieldRing(priceDirty, saved))}`}
-        />
-        {/* Billing cadence belongs with the price ("$1,499.00 /mo"), so its
-            select rides directly under the price input instead of claiming a
-            table column of its own. */}
-        <select
-          value={rec}
-          aria-label={t('quotes.editor.line.billingFrequencyAria')}
-          onChange={(e) => {
-            const next = e.target.value as QuoteLineRecurrence;
-            setRec(next); // optimistic — revert if the save fails
-            void edit({ recurrence: next }, 'rec').then((ok) => { if (!ok) setRec(line.recurrence); });
-          }}
-          disabled={fieldBusy('rec')}
-          data-testid={`quote-line-recurrence-${line.id}`}
-          className="ml-auto mt-1 block h-7 w-24 rounded-md border border-transparent bg-transparent py-0 pl-2 pr-6 text-xs text-muted-foreground transition-colors hover:border-border focus:border-border focus:outline-hidden disabled:opacity-60"
-        >
-          <option value="one_time">{t('quotes.editor.recurrence.one_time')}</option>
-          <option value="monthly">{t('quotes.editor.recurrence.monthly')}</option>
-          <option value="annual">{t('quotes.editor.recurrence.annual')}</option>
-        </select>
-      </td>
-      <td className="whitespace-nowrap px-1.5 py-2 text-right tabular-nums">
-        <span data-testid={`quote-line-total-${line.id}`}>{formatMoney(displayTotal, currency)}</span>
-        {rec !== 'one_time' && (
-          <span className="text-xs text-muted-foreground">{rec === 'monthly' ? t('quotes.editor.units.perMonth') : t('quotes.editor.units.perYear')}</span>
-        )}
-        <div className="text-xs font-normal text-muted-foreground" data-testid={`quote-line-tax-${line.id}`}>
-          {displayTax !== null
-            ? t('quotes.editor.table.taxSuffix', { amount: formatMoney(displayTax, currency) })
-            : taxable ? t('quotes.editor.table.taxable') : null}
-        </div>
-        <SrSaved show={saved} testId={`quote-line-saved-${line.id}`} />
-      </td>
-      <td className="px-1.5 py-2 text-right">
-        {/* ALL row-level actions live behind one overflow menu, so tabbing
-            through a line is data-entry only — one stop instead of four, and
-            no destructive button sitting mid-path between the price fields and
-            the description. Same menu grammar as the header kebab
-            (useMenuKeyboard: focus-on-open, arrow cycling, Esc → trigger). */}
-        <div ref={moveMenuRef} className="inline-block">
-          <button
-            ref={moveTriggerRef}
-            type="button"
-            onClick={(e) => {
-              if (movePos) { setMovePos(null); return; }
-              const r = e.currentTarget.getBoundingClientRect();
-              setMovePos({ top: r.bottom + 4, left: r.right });
-            }}
-            disabled={removeBusy}
-            aria-label={t('quotes.editor.actions.lineActions')}
-            aria-haspopup="menu"
-            aria-expanded={movePos !== null}
-            data-testid={`quote-line-actions-${line.id}`}
-            className="inline-flex h-7 w-7 items-center justify-center rounded text-muted-foreground hover:bg-muted disabled:opacity-30"
-          >
-            <MoreHorizontal className="h-4 w-4" aria-hidden />
-          </button>
-          {movePos && (
-            <div
-              role="menu"
-              ref={moveListRef}
-              onKeyDown={onMoveListKeyDown}
-              aria-label={t('quotes.editor.actions.lineActions')}
-              style={{ position: 'fixed', top: movePos.top, left: movePos.left, transform: 'translateX(-100%)' }}
-              className="z-50 w-max min-w-40 max-w-[min(20rem,calc(100vw-1rem))] rounded-md border bg-card py-1 shadow-md"
-              data-testid={`quote-line-actions-menu-${line.id}`}
-            >
-              <button
-                type="button"
-                role="menuitem"
-                tabIndex={-1}
-                disabled={isFirst}
-                onClick={() => { setMovePos(null); moveTriggerRef.current?.focus(); onMove(line, 'up'); }}
-                data-testid={`quote-line-move-up-${line.id}`}
-                className="block w-full px-3 py-1.5 text-left text-sm hover:bg-muted focus:bg-muted focus:outline-hidden disabled:opacity-40"
-              >
-                {t('quotes.editor.actions.moveLineUp')}
-              </button>
-              <button
-                type="button"
-                role="menuitem"
-                tabIndex={-1}
-                disabled={isLast}
-                onClick={() => { setMovePos(null); moveTriggerRef.current?.focus(); onMove(line, 'down'); }}
-                data-testid={`quote-line-move-down-${line.id}`}
-                className="block w-full px-3 py-1.5 text-left text-sm hover:bg-muted focus:bg-muted focus:outline-hidden disabled:opacity-40"
-              >
-                {t('quotes.editor.actions.moveLineDown')}
-              </button>
-              {moveTargets.length > 0 && !line.parentLineId && (
-                <>
-                  <p className="mt-1 border-t px-3 pb-0.5 pt-1.5 text-[11px] font-semibold uppercase tracking-wide text-muted-foreground">{t('quotes.editor.actions.moveTo')}</p>
-                  {moveTargets.map((t) => (
-                    <button
-                      key={t.id}
-                      type="button"
-                      role="menuitem"
-                      tabIndex={-1}
-                      title={t.label}
-                      onClick={() => { setMovePos(null); onMoveTo(line, t.id); }}
-                      data-testid={`quote-line-move-to-${line.id}-${t.id}`}
-                      className="block w-full truncate px-3 py-1.5 text-left text-sm hover:bg-muted focus:bg-muted focus:outline-hidden"
-                    >
-                      {t.label}
-                    </button>
-                  ))}
-                </>
-              )}
-              <div className="mt-1 border-t pt-1">
-                <button
-                  type="button"
-                  role="menuitem"
-                  tabIndex={-1}
-                  onClick={() => { setMovePos(null); onRemove(line); }}
-                  data-testid={`quote-line-remove-${line.id}`}
-                  className="block w-full px-3 py-1.5 text-left text-sm text-destructive hover:bg-destructive/10 focus:bg-destructive/10 focus:outline-hidden"
-                >
-                  {t('quotes.editor.actions.removeLine')}
-                </button>
-              </div>
-            </div>
-          )}
-        </div>
-      </td>
-    </tr>
-    {/* Full-width description row, so writers get a roomy, expandable box instead
-        of a cramped textarea squeezed into the narrow Description column. */}
-    <tr className="border-0" data-testid={`quote-line-desc-row-${line.id}`}>
-      <td colSpan={5} className="px-2 pb-2">
-        {/* Inline errors for the row's qty/price inputs — rendered full-width
-            directly under the row (the narrow cells above can't hold a message);
-            the offending input carries aria-invalid + a destructive ring. */}
-        {(fieldErrors.qty || fieldErrors.price) && (
-          <div className="mb-1 space-y-0.5">
-            {fieldErrors.qty && (
-              <p id={`quote-line-qty-error-${line.id}`} className="text-xs text-destructive" data-testid={`quote-line-qty-error-${line.id}`}>
-                {fieldErrors.qty}
-              </p>
-            )}
-            {fieldErrors.price && (
-              <p id={`quote-line-price-error-${line.id}`} className="text-xs text-destructive" data-testid={`quote-line-price-error-${line.id}`}>
-                {fieldErrors.price}
-              </p>
-            )}
-          </div>
-        )}
-        {descOpen && (
-          <textarea
-            ref={descRef}
-            value={desc}
-            aria-label={t('quotes.editor.line.descriptionAria')}
-            placeholder={t('quotes.editor.line.descriptionOptional')}
-            onChange={(e) => { setDesc(e.target.value); descEdited.current = true; autoGrowDesc(); }}
-            onBlur={commitDesc}
-            rows={2}
-            autoFocus={!(line.description ?? '').trim()}
-            disabled={fieldBusy('desc')}
-            data-testid={`quote-line-desc-${line.id}`}
-            className={`min-h-9 w-full resize-y overflow-hidden rounded-md border bg-transparent px-2 py-1 text-sm text-muted-foreground transition-colors focus:outline-hidden disabled:opacity-60 ${seamless(fieldRing(descDirty, saved))}`}
-          />
-        )}
-        <div className="mt-1 flex flex-wrap items-center gap-2">
-          {/* Taxable moved out of its own table column: it's an editing control,
-              not a per-glance figure (the computed tax shows under the Total). */}
-          <label className="flex items-center gap-1.5 text-xs text-muted-foreground">
-            <input
-              type="checkbox"
-              checked={taxable}
-              onChange={(e) => {
-                const next = e.target.checked;
-                setTaxable(next); // optimistic — revert if the save fails
-                void edit({ taxable: next }, 'taxable').then((ok) => { if (!ok) setTaxable(line.taxable); });
-              }}
-              disabled={fieldBusy('taxable')}
-              data-testid={`quote-line-taxable-${line.id}`}
-            />
-            {t('quotes.editor.table.taxable')}
-          </label>
-          {/* Deposit-eligible toggle appears only when the quote's deposit is
-              'selected_lines'. It's meaningful for one-time lines only (recurring
-              lines never count toward a deposit), so it's hidden for recurring rows. */}
-          {depositSelectMode && rec === 'one_time' && (
-            <label className="flex items-center gap-1.5 text-xs text-muted-foreground">
-              <input
-                type="checkbox"
-                checked={depositEligible}
-                onChange={(e) => {
-                  const next = e.target.checked;
-                  setDepositEligible(next); // optimistic — revert if the save fails
-                  void edit({ depositEligible: next }, 'deposit').then((ok) => { if (!ok) setDepositEligible(line.depositEligible ?? false); });
-                }}
-                disabled={fieldBusy('deposit')}
-                data-testid={`line-deposit-eligible-${line.id}`}
-              />
-              {t('quotes.editor.deposit.eligibleAria')}
-            </label>
-          )}
-          {!descOpen && (
-            <button
-              type="button"
-              onClick={() => setDescOpen(true)}
-              data-testid={`quote-line-desc-open-${line.id}`}
-              className="inline-flex h-8 items-center rounded-md border border-transparent px-2 text-xs font-medium text-muted-foreground hover:border-border hover:text-foreground"
-            >
-              <span aria-hidden="true">+</span>&nbsp;{t('quotes.editor.line.addDescription')}
-            </button>
-          )}
-          {descOpen && (name.trim() || desc.trim()) && (
-            <PolishButton
-              disabled={fieldBusy('polish')}
-              idSuffix={`quote-line-${line.id}`}
-              compact
-              getText={() => ({ name, description: desc })}
-              onApply={(r) => {
-                const patch: { name?: string | null; description?: string | null } = {};
-                if (r.name !== null) { setName(r.name); nameEdited.current = false; patch.name = r.name || null; }
-                if (r.description !== null) { setDesc(r.description); descEdited.current = false; patch.description = r.description || null; }
-                if (Object.keys(patch).length) void edit(patch, 'polish');
-              }}
-            />
-          )}
-          {/* Per-line product image controls. The thumbnail itself renders next
-              to the name; these manage it. Same 5MB/type limits as image blocks. */}
-          <input
-            ref={imageInputRef}
-            type="file"
-            accept="image/png,image/jpeg,image/webp"
-            className="hidden"
-            data-testid={`quote-line-image-input-${line.id}`}
-            onChange={(e) => {
-              const f = e.target.files?.[0];
-              e.target.value = ''; // allow re-picking the same file
-              if (f) attachImage(f);
-            }}
-          />
-          <button
-            type="button"
-            onClick={() => imageInputRef.current?.click()}
-            disabled={imageBusy || fieldBusy('image')}
-            aria-busy={imageBusy}
-            data-testid={`quote-line-image-attach-${line.id}`}
-            className="inline-flex h-8 items-center gap-1 rounded-md border px-2 text-xs font-medium hover:bg-muted disabled:opacity-50"
-          >
-            {imageBusy && <Loader2 className="h-3 w-3 animate-spin" aria-hidden="true" />}
-            {imageBusy ? t('quotes.editor.actions.uploading') : line.imageId ? t('quotes.editor.actions.replaceImage') : t('quotes.editor.actions.addImage')}
-          </button>
-          <button
-            type="button"
-            onClick={() => setImageUrlOpen((v) => !v)}
-            disabled={imageBusy || fieldBusy('image')}
-            aria-expanded={imageUrlOpen}
-            data-testid={`quote-line-image-url-toggle-${line.id}`}
-            className="inline-flex h-8 items-center rounded-md border px-2 text-xs font-medium hover:bg-muted disabled:opacity-50"
-          >
-            {t('quotes.editor.actions.fromUrl')}
-          </button>
-          {line.imageId && !imageBusy && (
-            <button
-              type="button"
-              onClick={() => void edit({ imageId: null }, 'image')}
-              disabled={fieldBusy('image')}
-              data-testid={`quote-line-image-remove-${line.id}`}
-              className="inline-flex h-8 items-center rounded-md border px-2 text-xs font-medium text-muted-foreground hover:bg-muted disabled:opacity-50"
-            >
-              {t('quotes.editor.actions.removeImage')}
-            </button>
-          )}
-          {/* URL disclosure: w-full forces it onto its own row under the parent's
-              flex-wrap so the input has room. Server copies the bytes in
-              (SSRF-guarded); Fetch is disabled until a URL is entered, matching
-              the image block's submit gate. */}
-          {imageUrlOpen && (
-            <div className="flex w-full items-center gap-2">
-              <input
-                type="url"
-                value={imageUrlDraft}
-                onChange={(e) => setImageUrlDraft(e.target.value)}
-                placeholder={t('quotes.editor.addSection.imageUrlPlaceholder')}
-                disabled={imageBusy}
-                aria-label={t('quotes.editor.line.imageUrlAria')}
-                data-testid={`quote-line-image-url-input-${line.id}`}
-                className="h-8 min-w-0 flex-1 rounded-md border bg-background px-2 text-xs focus:outline-hidden focus:ring-2 focus:ring-ring disabled:opacity-60"
-              />
-              <button
-                type="button"
-                onClick={attachImageFromUrl}
-                disabled={imageBusy || fieldBusy('image') || !imageUrlDraft.trim()}
-                data-testid={`quote-line-image-url-fetch-${line.id}`}
-                className="inline-flex h-8 items-center rounded-md bg-primary px-2 text-xs font-medium text-primary-foreground hover:opacity-90 disabled:opacity-50"
-              >
-                {t('quotes.editor.actions.fetch')}
-              </button>
-              <button
-                type="button"
-                onClick={() => { setImageUrlOpen(false); setImageUrlDraft(''); }}
-                disabled={imageBusy}
-                data-testid={`quote-line-image-url-cancel-${line.id}`}
-                className="inline-flex h-8 items-center rounded-md border px-2 text-xs font-medium text-muted-foreground hover:bg-muted disabled:opacity-50"
-              >
-                {t('quotes.editor.actions.cancel')}
-              </button>
-            </div>
-          )}
-        </div>
-      </td>
-    </tr>
-    {/* Internal-only cost/markup/profit band — never shown to the customer.
-        Collapsed by default via the editor's "Show cost & margin" toggle; kept in
-        the DOM (hidden) rather than unmounted so totals/draft wiring stays live. */}
-    <tr className={`border-0 ${showInternal ? '' : 'hidden'}`} data-testid={`quote-line-internal-${line.id}`}>
-      <td colSpan={5} className="px-2 pb-2">
-        <div className="flex flex-wrap items-center gap-x-3 gap-y-1 rounded-md bg-muted/40 px-2 py-1 text-xs text-[hsl(220_12%_40%)] dark:text-muted-foreground">
-          {/* Full disclaimer on the first row; a subtle "Internal" tag persists on
-              every following row so a writer scanning mid-table never mistakes the
-              cost/markup band for customer-facing copy. */}
-          <span className="font-medium uppercase tracking-wide">{isFirst ? t('quotes.editor.internal.full') : t('quotes.editor.internal.short')}</span>
-          <label className="flex items-center gap-1">{t('quotes.editor.line.sku')}
-            <input
-              type="text"
-              value={sku}
-              onChange={(e) => { setSku(e.target.value); skuEdited.current = true; }}
-              onBlur={commitSku}
-              disabled={fieldBusy('sku')}
-              data-testid={`quote-line-sku-${line.id}`}
-              className={`h-6 w-28 rounded border bg-background px-1 text-foreground transition-shadow ${fieldRing(skuDirty, saved)}`}
-            />
-          </label>
-          <label className="flex items-center gap-1">{t('quotes.editor.line.partNumberAbbr')}
-            <input
-              type="text"
-              value={partNumber}
-              onChange={(e) => { setPartNumber(e.target.value); partEdited.current = true; }}
-              onBlur={commitPartNumber}
-              disabled={fieldBusy('pn')}
-              data-testid={`quote-line-partnumber-${line.id}`}
-              className={`h-6 w-28 rounded border bg-background px-1 text-foreground transition-shadow ${fieldRing(partDirty, saved)}`}
-            />
-          </label>
-          <label className="flex items-center gap-1">{t('quotes.editor.line.cost')}
-            <input
-              type="number" min="0" step="0.01"
-              value={cost}
-              onChange={(e) => { setCost(e.target.value); costEdited.current = true; setFieldError('cost', null); }}
-              onBlur={commitCost}
-              disabled={fieldBusy('cost')}
-              aria-invalid={fieldErrors.cost ? true : undefined}
-              aria-describedby={fieldErrors.cost ? `quote-line-cost-error-${line.id}` : undefined}
-              data-testid={`quote-line-cost-${line.id}`}
-              className={`h-6 w-20 rounded border bg-background px-1 text-right tabular-nums text-foreground transition-shadow ${fieldErrors.cost ? 'border-destructive ring-1 ring-destructive' : fieldRing(costDirty, saved)}`}
-            />
-          </label>
-          {fieldErrors.cost && (
-            <p id={`quote-line-cost-error-${line.id}`} className="w-full text-xs font-normal normal-case tracking-normal text-destructive" data-testid={`quote-line-cost-error-${line.id}`}>
-              {fieldErrors.cost}
-            </p>
-          )}
-          <label className="flex items-center gap-1">{t('quotes.editor.line.markup')}
-            <input
-              type="number" step="0.1"
-              value={markupInput}
-              onFocus={() => { markupFocused.current = true; }}
-              onChange={(e) => setMarkupInput(e.target.value)}
-              onBlur={(e) => { markupFocused.current = false; onMarkupCommit(e.target.value); }}
-              disabled={fieldBusy('price') || cost.trim() === ''}
-              // Tell keyboard/SR users WHY the field is disabled — sighted users can
-              // see the empty cost field, AT users can't.
-              title={cost.trim() === '' ? t('quotes.editor.line.enterCostFirstMarkup') : undefined}
-              aria-describedby={cost.trim() === '' ? `quote-line-markup-hint-${line.id}` : undefined}
-              data-testid={`quote-line-markup-${line.id}`}
-              className="h-6 w-16 rounded border bg-background px-1 text-right tabular-nums text-foreground disabled:opacity-60"
-            />{t('quotes.editor.symbols.percent')}
-            {cost.trim() === '' && <span id={`quote-line-markup-hint-${line.id}`} className="sr-only">{t('quotes.editor.line.enterCostFirstMarkupSentence')}</span>}
-          </label>
-          <span className="ml-auto">{t('quotes.editor.line.profit')}{' '}
-            <span className="font-medium tabular-nums text-foreground" data-testid={`quote-line-net-${line.id}`}>
-              {netCents === null ? t('quotes.editor.symbols.notAvailable') : formatMoney(fromCents(netCents), currency)}
-            </span>
-          </span>
-        </div>
-      </td>
-    </tr>
-    </>
-  );
-}
-
-// Small product thumbnail for a catalog-sourced quote line. GET /catalog/:id/image
-// needs the Bearer header (a bare <img src> would 401), and 404s when the item has
-// no image — so we fetchWithAuth → blob → object URL and render nothing on miss.
-function CatalogLineThumb({ catalogItemId }: { catalogItemId: string }) {
-  const [url, setUrl] = useState<string>();
-  useEffect(() => {
-    let objectUrl: string | undefined;
-    let cancelled = false;
-    void (async () => {
-      try {
-        const res = await fetchWithAuth(catalogItemImagePath(catalogItemId));
-        if (!res.ok) return; // 404 = no image; render nothing
-        const blob = await res.blob();
-        if (cancelled) return;
-        objectUrl = window.URL.createObjectURL(blob);
-        setUrl(objectUrl);
-      } catch {
-        // no image / load failure — render nothing
-      }
-    })();
-    return () => { cancelled = true; if (objectUrl) window.URL.revokeObjectURL(objectUrl); };
-  }, [catalogItemId]);
-
-  if (!url) return null;
-  return <img src={url} alt="" className="h-10 w-10 shrink-0 rounded border object-contain" data-testid="quote-line-thumb" />;
-}
-
-// Per-line uploaded image thumbnail (GET /quotes/:id/images/:imageId needs the
-// Bearer header — same contract as CatalogLineThumb: render nothing on miss).
-function LineImageThumb({ quoteId, imageId }: { quoteId: string; imageId: string }) {
-  const { url } = useAuthedImage(quoteImageUrl(quoteId, imageId));
-  if (!url) return null;
-  return <img src={url} alt="" className="h-10 w-10 shrink-0 rounded border object-contain" data-testid="quote-line-image-thumb" />;
-}
-
-// A persisted `contract` block in the editor. The admin serialization attaches
-// the raw authoring fields (content.authoring), so — unlike portal/public — the
-// editor can render an editable manual-variable form (PATCH variableValues), an
-// explicit "Update to vN" nudge when the pin is behind the latest published
-// version, and an inline list of unresolved (empty) manual variables (the
-// send-time CONTRACT_VARIABLES_UNRESOLVED equivalent surfaced on the block).
-// Without authoring (legacy / uploaded / read-only user) it degrades to a
-// read-only summary card.
-function ContractBlockEditor({
-  block, canWrite, onEditBlock,
-}: {
-  block: QuoteBlock;
-  canWrite: boolean;
-  onEditBlock: (block: QuoteBlock, content: Record<string, unknown>) => Promise<boolean>;
-}) {
-  const { t } = useTranslation('billing');
-  const c = (block.content ?? {}) as Partial<ContractBlockContent>;
-  const authoring = c.authoring;
-  const templateName = c.templateName?.trim() || t('quotes.editor.contract.untitledTemplate');
-  const versionNumber = c.versionNumber ?? 0;
-
-  const manualVars = useMemo(() => authoring?.declaredVariables.filter((v) => v.kind === 'manual') ?? [], [authoring]);
-  const autoVars = useMemo(() => authoring?.declaredVariables.filter((v) => v.kind === 'auto') ?? [], [authoring]);
-
-  const [draft, setDraft] = useState<Record<string, string>>(() => ({ ...(authoring?.variableValues ?? {}) }));
-  const [errors, setErrors] = useState<Record<string, string>>({});
-  const [busy, setBusy] = useState(false);
-  const [saved, setSaved] = useState(false);
-  const savedTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
-  useEffect(() => () => { if (savedTimer.current) clearTimeout(savedTimer.current); }, []);
-  const flash = useCallback(() => {
-    setSaved(true);
-    if (savedTimer.current) clearTimeout(savedTimer.current);
-    savedTimer.current = setTimeout(() => setSaved(false), 1500);
-  }, []);
-
-  // Resync the draft from the server's persisted values after a save's refetch,
-  // but only when the user hasn't diverged (same guard as the heading/rich-text
-  // drafts): if the local draft no longer matches what we last synced, keep it.
-  const lastSynced = useRef(JSON.stringify(authoring?.variableValues ?? {}));
-  useEffect(() => {
-    const nextStr = JSON.stringify(authoring?.variableValues ?? {});
-    setDraft((cur) => (JSON.stringify(cur) === lastSynced.current ? { ...(authoring?.variableValues ?? {}) } : cur));
-    lastSynced.current = nextStr;
-  }, [authoring]);
-
-  const unfilled = useMemo(
-    () => manualVars.filter((v) => !(draft[v.name] ?? '').trim()).map((v) => v.name),
-    [manualVars, draft],
-  );
-  const latestNumber = authoring?.latestPublishedVersionNumber ?? null;
-  const latestId = authoring?.latestPublishedVersionId ?? null;
-  const canUpdate = latestId != null && latestNumber != null && latestNumber > versionNumber;
-
-  const commit = useCallback(async (versionOverride?: string) => {
-    if (!authoring) return;
-    const names = manualVars.map((v) => v.name);
-    const missing = names.filter((n) => !(draft[n] ?? '').trim());
-    if (missing.length > 0) {
-      setErrors(Object.fromEntries(missing.map((n) => [n, t('quotes.editor.contract.variableRequired')])));
-      return;
-    }
-    setErrors({});
-    const variableValues = Object.fromEntries(names.map((n) => [n, (draft[n] ?? '').trim()]));
-    const content: Record<string, unknown> = {
-      templateId: authoring.templateId,
-      templateVersionId: versionOverride ?? authoring.templateVersionId,
-      variableValues,
-      ...(c.label?.trim() ? { label: c.label.trim() } : {}),
-    };
-    setBusy(true);
-    try { if (await onEditBlock(block, content)) flash(); } finally { setBusy(false); }
-  }, [authoring, manualVars, draft, c.label, block, onEditBlock, flash, t]);
-
-  // No authoring (legacy/uploaded) or read-only user → read-only summary card.
-  if (!authoring || !canWrite) {
-    return (
-      <div className="space-y-2 text-sm" data-testid={`quote-block-contract-content-${block.id}`}>
-        <div className="flex flex-wrap items-center gap-2">
-          <span className="font-medium">{c.label?.trim() || templateName}</span>
-          <span className="rounded-full border px-2 py-0.5 text-xs text-muted-foreground">
-            {t('quotes.editor.contract.pinnedVersion', { version: versionNumber })}
-          </span>
-        </div>
-        {c.label?.trim() && <p className="text-xs text-muted-foreground">{templateName}</p>}
-        {!authoring && <p className="text-xs text-muted-foreground">{t('quotes.editor.contract.readOnlyHint')}</p>}
-      </div>
-    );
-  }
-
-  return (
-    <div className="space-y-3 text-sm" data-testid={`quote-block-contract-editor-${block.id}`}>
-      <div className="flex flex-wrap items-center gap-2">
-        <span className="font-medium">{c.label?.trim() || templateName}</span>
-        <span className="rounded-full border px-2 py-0.5 text-xs text-muted-foreground" data-testid={`quote-block-contract-version-${block.id}`}>
-          {t('quotes.editor.contract.pinnedVersion', { version: versionNumber })}
-        </span>
-        <SrSaved show={saved} testId={`quote-block-contract-saved-${block.id}`} />
-        {canUpdate && (
-          <button
-            type="button"
-            onClick={() => void commit(latestId!)}
-            disabled={busy}
-            data-testid={`quote-block-contract-update-${block.id}`}
-            className="ml-auto rounded-md border border-primary/40 bg-primary/10 px-2 py-0.5 text-xs font-medium text-primary hover:bg-primary/20 disabled:opacity-50"
-          >
-            {t('quotes.editor.contract.updateToVersion', { version: latestNumber })}
-          </button>
-        )}
-      </div>
-
-      {autoVars.length > 0 && (
-        <div>
-          <p className="mb-1 text-xs font-medium text-muted-foreground">{t('quotes.editor.contract.autoVariablesTitle')}</p>
-          <ul className="space-y-1">
-            {autoVars.map((v) => (
-              <li
-                key={v.name}
-                data-testid={`quote-block-contract-auto-${block.id}-${v.name}`}
-                className="flex items-center justify-between rounded-md border bg-muted/40 px-2 py-1 text-xs"
-              >
-                <span className="font-medium">{v.label ?? v.name}</span>
-                <span className="font-mono text-muted-foreground">{`{{${v.name}}}`}</span>
-              </li>
-            ))}
-          </ul>
-        </div>
-      )}
-
-      {manualVars.length > 0 && (
-        <div className="space-y-2">
-          <p className="text-xs font-medium text-muted-foreground">{t('quotes.editor.contract.manualVariablesTitle')}</p>
-          {manualVars.map((v) => (
-            <div key={v.name}>
-              <label htmlFor={`quote-block-contract-var-${block.id}-${v.name}`} className="mb-0.5 block text-xs text-muted-foreground">
-                {v.label ?? v.name}
-              </label>
-              <input
-                id={`quote-block-contract-var-${block.id}-${v.name}`}
-                type="text"
-                value={draft[v.name] ?? ''}
-                onChange={(e) => {
-                  const val = e.target.value;
-                  setDraft((cur) => ({ ...cur, [v.name]: val }));
-                  setErrors((cur) => { if (!cur[v.name]) return cur; const next = { ...cur }; delete next[v.name]; return next; });
-                }}
-                disabled={busy}
-                data-testid={`quote-block-contract-var-${block.id}-${v.name}`}
-                aria-invalid={errors[v.name] ? true : undefined}
-                className={`h-9 w-full rounded-md border bg-background px-3 text-sm focus:outline-hidden focus:ring-2 focus:ring-ring disabled:opacity-60 ${errors[v.name] ? 'border-destructive' : ''}`}
-              />
-              {errors[v.name] && (
-                <p className="mt-0.5 text-xs text-destructive" data-testid={`quote-block-contract-var-error-${block.id}-${v.name}`}>
-                  {errors[v.name]}
-                </p>
-              )}
-            </div>
-          ))}
-        </div>
-      )}
-
-      {unfilled.length > 0 && (
-        <p className="rounded-md border border-warning/40 bg-warning/10 px-2 py-1 text-xs text-warning-foreground dark:text-warning" data-testid={`quote-block-contract-unresolved-${block.id}`}>
-          {t('quotes.editor.contract.unresolvedWarning', { names: unfilled.join(', ') })}
-        </p>
-      )}
-
-      {manualVars.length > 0 && (
-        <div className="flex justify-end">
-          <button
-            type="button"
-            onClick={() => void commit()}
-            disabled={busy}
-            data-testid={`quote-block-contract-save-${block.id}`}
-            className="inline-flex h-8 items-center justify-center rounded-md bg-primary px-3 text-xs font-medium text-primary-foreground hover:opacity-90 disabled:opacity-50"
-          >
-            {t('quotes.editor.contract.saveVariables')}
-          </button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Editor image preview. GET /quotes/:id/images/:imageId requires the Bearer auth
-// header, so a bare <img src> would 401 (web-1). Mirror QuoteWorkspace's PDF
-// preview: fetchWithAuth → blob → object URL, revoked on unmount/change.
-function QuoteImagePreview({ quoteId, imageId, caption }: { quoteId: string; imageId: string; caption?: string }) {
-  const { t } = useTranslation('billing');
-  const [url, setUrl] = useState<string>();
-  const [failed, setFailed] = useState(false);
-
-  useEffect(() => {
-    let objectUrl: string | undefined;
-    let cancelled = false;
-    void (async () => {
-      try {
-        const res = await fetchWithAuth(quoteImageUrl(quoteId, imageId));
-        if (!res.ok) { if (!cancelled) setFailed(true); return; }
-        const blob = await res.blob();
-        if (cancelled) return;
-        objectUrl = window.URL.createObjectURL(blob);
-        setUrl(objectUrl);
-      } catch {
-        if (!cancelled) setFailed(true);
-      }
-    })();
-    return () => { cancelled = true; if (objectUrl) window.URL.revokeObjectURL(objectUrl); };
-  }, [quoteId, imageId]);
-
-  if (failed) return <p className="text-sm text-muted-foreground">{t('quotes.editor.image.previewUnavailable')}</p>;
-  if (!url) return <div className="h-24 w-full animate-pulse rounded border bg-muted" data-testid="quote-image-loading" />;
-  return <img src={url} alt={caption || t('quotes.editor.image.alt')} className="max-h-64 rounded border" />;
 }
